refactor(rel): extract temporal Arrow and UDF adapters
