+++ crates/graphforge-rel/src/expr.rs
@@ -5,6 +5,20 @@
 //! that can be consumed by operator lowering (#575, #576).
 
 mod aggregates;
+mod temporal_udfs;
+use temporal_udfs::{
+    CYPHER_DATE_COMPONENT, CYPHER_DATE_PROJECT, CYPHER_DATE_TRUNCATE, CYPHER_DATETIME_PROJECT,
+    CYPHER_DATETIME_TRUNCATE, CYPHER_DURATION_ADD, CYPHER_DURATION_BETWEEN,
+    CYPHER_DURATION_COMPONENT, CYPHER_DURATION_PARSE, CYPHER_DURATION_SCALE,
+    CYPHER_LOCALDATETIME_PROJECT, CYPHER_LOCALDATETIME_TRUNCATE, CYPHER_LOCALTIME_PROJECT,
+    CYPHER_LOCALTIME_TRUNCATE, CYPHER_TEMPORAL_ARITH, CYPHER_TEMPORAL_COMPONENT,
+    CYPHER_TEMPORAL_ZONE_STR, CYPHER_TIME_PROJECT, CYPHER_TIME_TRUNCATE, date_scalar,
+    date_struct_value, datetime_scalar, datetime_struct_parts, dur_secs_nanos, duration_scalar,
+    duration_struct_parts, duration_value_to_ir, is_date_struct, is_datetime_struct,
+    is_duration_struct, is_localdatetime_struct, is_temporal_clock_fn, is_time_struct,
+    localdatetime_scalar, localdatetime_struct_parts, temporal_accessor_valid,
+    temporal_null_scalar, time_scalar, time_struct_parts,
+};
 mod value_semantics;
 
 pub use value_semantics::decode_het_scalar;
@@ -6978,817 +6992,6 @@ pub(crate) fn is_het_struct_type(t: Option<&DataType>) -> bool {
     matches!(t, Some(DataType::Struct(fields)) if fields.iter().any(|f| f.name() == graphforge_value::heterogeneous::TAG))
 }
 
-/// `date` component accessor (`Temporal5`): `cypher_date_component(date, name)`
-/// where `date` is a `Date32` and `name` is the accessor (`year`/`quarter`/
-/// `month`/`week`/`weekYear`/`day`/`ordinalDay`/`weekDay`/`dayOfQuarter`).
-/// Returns the component as `Int64`. (ADR 0009 / #920)
-static CYPHER_DATE_COMPONENT: LazyLock<ScalarUDF> =
-    LazyLock::new(|| ScalarUDF::new_from_impl(CypherDateComponent::new()));
-
-#[derive(Debug, PartialEq, Eq, Hash)]
-struct CypherDateComponent {
-    signature: Signature,
-}
-
-impl CypherDateComponent {
-    fn new() -> Self {
-        Self {
-            signature: Signature::any(2, Volatility::Immutable),
-        }
-    }
-}
-
-impl ScalarUDFImpl for CypherDateComponent {
-    fn name(&self) -> &'static str {
-        "cypher_date_component"
-    }
-
-    fn signature(&self) -> &Signature {
-        &self.signature
-    }
-
-    fn return_type(&self, _arg_types: &[DataType]) -> datafusion::error::Result<DataType> {
-        Ok(DataType::Int64)
-    }
-
-    fn invoke_with_args(
-        &self,
-        args: ScalarFunctionArgs,
-    ) -> datafusion::error::Result<ColumnarValue> {
-        use crate::temporal::date_component;
-        use datafusion::arrow::array::{Array, Int64Array, StringArray, StructArray};
-
-        validate_heterogeneous_arguments(&args.args)?;
-        let rows = args.number_rows;
-        let dates = args.args[0].to_array(rows)?;
-        let names = args.args[1].to_array(rows)?;
-        let d = dates.as_any().downcast_ref::<StructArray>();
-        let n = names.as_any().downcast_ref::<StringArray>();
-        let out: Int64Array = (0..rows)
-            .map(|i| {
-                let (d, n) = (d?, n?);
-                if n.is_null(i) {
-                    return None;
-                }
-                date_component(date_struct_value(d, i)?, n.value(i))
-            })
-            .collect();
-        Ok(ColumnarValue::Array(std::sync::Arc::new(out)))
-    }
-}
-
-// ---------------------------------------------------------------------------
-// cypher_duration_component UDF
-// ---------------------------------------------------------------------------
-
-/// `duration` component accessor (`d.days`/`d.seconds`/`d.monthsOfQuarter`/…):
-/// `[interval_value, component_name]` → `Int64`. (#920)
-static CYPHER_DURATION_COMPONENT: LazyLock<ScalarUDF> =
-    LazyLock::new(|| ScalarUDF::new_from_impl(CypherDurationComponent::new()));
-
-#[derive(Debug, PartialEq, Eq, Hash)]
-struct CypherDurationComponent {
-    signature: Signature,
-}
-
-impl CypherDurationComponent {
-    fn new() -> Self {
-        Self {
-            signature: Signature::any(2, Volatility::Immutable),
-        }
-    }
-}
-
-impl ScalarUDFImpl for CypherDurationComponent {
-    fn name(&self) -> &'static str {
-        "cypher_duration_component"
-    }
-
-    fn signature(&self) -> &Signature {
-        &self.signature
-    }
-
-    fn return_type(&self, _arg_types: &[DataType]) -> datafusion::error::Result<DataType> {
-        Ok(DataType::Int64)
-    }
-
-    fn invoke_with_args(
-        &self,
-        args: ScalarFunctionArgs,
-    ) -> datafusion::error::Result<ColumnarValue> {
-        use crate::temporal::duration_component;
-        use datafusion::arrow::array::{Array, Int64Array, StringArray, StructArray};
-
-        validate_heterogeneous_arguments(&args.args)?;
-        let rows = args.number_rows;
-        let durs = args.args[0].to_array(rows)?;
-        let names = args.args[1].to_array(rows)?;
-        let d = durs.as_any().downcast_ref::<StructArray>();
-        let n = names.as_any().downcast_ref::<StringArray>();
-        let out: Int64Array = (0..rows)
-            .map(|i| {
-                let (d, n) = (d?, n?);
-                if d.is_null(i) || n.is_null(i) {
-                    return None;
-                }
-                duration_component(&duration_struct_parts(d, i)?, n.value(i))
-            })
-            .collect();
-        Ok(ColumnarValue::Array(std::sync::Arc::new(out)))
-    }
-}
-
-// ---------------------------------------------------------------------------
-// cypher_temporal_component / cypher_temporal_zone_str UDFs (Temporal5 accessors)
-// ---------------------------------------------------------------------------
-
-/// Whether `name` is a valid component accessor for a typed-temporal COLUMN type
-/// other than `Date32`/duration (which dispatch separately): `localtime`
-/// (`Time64`), or a `time`/`localdatetime`/`datetime` struct. Gates the accessor
-/// dispatch so a non-temporal property access still falls through to a column
-/// lookup. (#1008)
-fn temporal_accessor_valid(dt: &DataType, name: &str) -> bool {
-    use crate::temporal::{
-        is_date_accessor, is_epoch_accessor, is_time_accessor, is_zone_int_accessor,
-        is_zone_str_accessor,
-    };
-    match dt {
-        DataType::Time64(_) => is_time_accessor(name),
-        DataType::Struct(_) if is_time_struct(dt) => {
-            is_time_accessor(name) || is_zone_int_accessor(name) || is_zone_str_accessor(name)
-        }
-        DataType::Struct(_) if is_localdatetime_struct(dt) => {
-            is_date_accessor(name) || is_time_accessor(name)
-        }
-        DataType::Struct(_) if is_datetime_struct(dt) => {
-            is_date_accessor(name)
-                || is_time_accessor(name)
-                || is_zone_int_accessor(name)
-                || is_zone_str_accessor(name)
-                || is_epoch_accessor(name)
-        }
-        _ => false,
-    }
-}
-
-/// `Temporal5` INT component accessor (`d.hour`/`d.year`/`d.offsetSeconds`/
-/// `d.epochMillis`/…): `[value, name]` → `Int64`. `value` is a typed `localtime`
-/// (`Time64`) or `time`/`localdatetime`/`datetime` struct; the UDF inspects the
-/// Arrow type to extract the relevant field. (#1008)
-static CYPHER_TEMPORAL_COMPONENT: LazyLock<ScalarUDF> =
-    LazyLock::new(|| ScalarUDF::new_from_impl(CypherTemporalComponent::new()));
-
-#[derive(Debug, PartialEq, Eq, Hash)]
-struct CypherTemporalComponent {
-    signature: Signature,
-}
-
-impl CypherTemporalComponent {
-    fn new() -> Self {
-        Self {
-            signature: Signature::any(2, Volatility::Immutable),
-        }
-    }
-}
-
-impl ScalarUDFImpl for CypherTemporalComponent {
-    fn name(&self) -> &'static str {
-        "cypher_temporal_component"
-    }
-
-    fn signature(&self) -> &Signature {
-        &self.signature
-    }
-
-    fn return_type(&self, _arg_types: &[DataType]) -> datafusion::error::Result<DataType> {
-        Ok(DataType::Int64)
-    }
-
-    fn invoke_with_args(
-        &self,
-        args: ScalarFunctionArgs,
-    ) -> datafusion::error::Result<ColumnarValue> {
-        use crate::temporal::{
-            date_component, epoch_component, is_date_accessor, is_time_accessor,
-            is_zone_int_accessor, time_component, zone_int_component,
-        };
-        use datafusion::arrow::array::{
-            Array, Int64Array, StringArray, StructArray, Time64NanosecondArray,
-        };
-        use datafusion::arrow::datatypes::TimeUnit;
-
-        validate_heterogeneous_arguments(&args.args)?;
-        let rows = args.number_rows;
-        let vals = args.args[0].to_array(rows)?;
-        let names = args.args[1].to_array(rows)?;
-        let n = names.as_any().downcast_ref::<StringArray>();
-        let out: Int64Array = (0..rows)
-            .map(|i| {
-                let n = n?;
-                if vals.is_null(i) || n.is_null(i) {
-                    return None;
-                }
-                let name = n.value(i);
-                match vals.data_type() {
-                    DataType::Time64(TimeUnit::Nanosecond) => {
-                        let v = vals.as_any().downcast_ref::<Time64NanosecondArray>()?;
-                        time_component(v.value(i), name)
-                    }
-                    DataType::Struct(_) => {
-                        let s = vals.as_any().downcast_ref::<StructArray>()?;
-                        if is_time_struct(vals.data_type()) {
-                            let (nanos, offset) = time_struct_parts(s, i)?;
-                            if is_zone_int_accessor(name) {
-                                zone_int_component(offset, name)
-                            } else {
-                                time_component(nanos, name)
-                            }
-                        } else if is_localdatetime_struct(vals.data_type()) {
-                            let (days, nanos) = localdatetime_struct_parts(s, i)?;
-                            if is_date_accessor(name) {
-                                date_component(days, name)
-                            } else {
-                                time_component(nanos, name)
-                            }
-                        } else if is_datetime_struct(vals.data_type()) {
-                            let (days, nanos, offset, _) = datetime_struct_parts(s, i)?;
-                            if is_date_accessor(name) {
-                                date_component(days, name)
-                            } else if is_time_accessor(name) {
-                                time_component(nanos, name)
-                            } else if is_zone_int_accessor(name) {
-                                zone_int_component(offset, name)
-                            } else {
-                                epoch_component(days, nanos, offset, name)
-                            }
-                        } else {
-                            None
-                        }
-                    }
-                    _ => None,
-                }
-            })
-            .collect();
-        Ok(ColumnarValue::Array(std::sync::Arc::new(out)))
-    }
-}
-
-/// `Temporal5` STRING component accessor (`d.timezone`/`d.offset`): `[value,
-/// name]` → `Utf8`. `value` is a typed `time` or `datetime` struct. (#1008)
-static CYPHER_TEMPORAL_ZONE_STR: LazyLock<ScalarUDF> =
-    LazyLock::new(|| ScalarUDF::new_from_impl(CypherTemporalZoneStr::new()));
-
-#[derive(Debug, PartialEq, Eq, Hash)]
-struct CypherTemporalZoneStr {
-    signature: Signature,
-}
-
-impl CypherTemporalZoneStr {
-    fn new() -> Self {
-        Self {
-            signature: Signature::any(2, Volatility::Immutable),
-        }
-    }
-}
-
-impl ScalarUDFImpl for CypherTemporalZoneStr {
-    fn name(&self) -> &'static str {
-        "cypher_temporal_zone_str"
-    }
-
-    fn signature(&self) -> &Signature {
-        &self.signature
-    }
-
-    fn return_type(&self, _arg_types: &[DataType]) -> datafusion::error::Result<DataType> {
-        Ok(DataType::Utf8)
-    }
-
-    fn invoke_with_args(
-        &self,
-        args: ScalarFunctionArgs,
-    ) -> datafusion::error::Result<ColumnarValue> {
-        use crate::temporal::zone_str_component;
-        use datafusion::arrow::array::{Array, StringArray, StructArray};
-
-        validate_heterogeneous_arguments(&args.args)?;
-        let rows = args.number_rows;
-        let vals = args.args[0].to_array(rows)?;
-        let names = args.args[1].to_array(rows)?;
-        let n = names.as_any().downcast_ref::<StringArray>();
-        let out: StringArray = (0..rows)
-            .map(|i| {
-                let n = n?;
-                if vals.is_null(i) || n.is_null(i) {
-                    return None;
-                }
-                let name = n.value(i);
-                let s = vals.as_any().downcast_ref::<StructArray>()?;
-                if is_time_struct(vals.data_type()) {
-                    let (_, offset) = time_struct_parts(s, i)?;
-                    zone_str_component(offset, None, name)
-                } else if is_datetime_struct(vals.data_type()) {
-                    let (_, _, offset, zone) = datetime_struct_parts(s, i)?;
-                    zone_str_component(offset, zone.as_deref(), name)
-                } else {
-                    None
-                }
-            })
-            .collect();
-        Ok(ColumnarValue::Array(std::sync::Arc::new(out)))
-    }
-}
-
-// ---------------------------------------------------------------------------
-// cypher_duration_between UDF
-// ---------------------------------------------------------------------------
-
-/// `duration.between(a, b)` / `inMonths` / `inDays` / `inSeconds` (`Temporal10`):
-/// `[a, b, mode]` → `Interval(MonthDayNano)`. `a`/`b` are typed temporals
-/// (`Date32`/`Time64`/`localdatetime`/`time`/`datetime` struct); `mode` selects
-/// the family member. (#920)
-static CYPHER_DURATION_BETWEEN: LazyLock<ScalarUDF> =
-    LazyLock::new(|| ScalarUDF::new_from_impl(CypherDurationBetween::new()));
-
-#[derive(Debug, PartialEq, Eq, Hash)]
-struct CypherDurationBetween {
-    signature: Signature,
-}
-
-impl CypherDurationBetween {
-    fn new() -> Self {
-        Self {
-            signature: Signature::any(3, Volatility::Immutable),
-        }
-    }
-}
-
-/// Extract a [`BetweenOperand`](crate::temporal::BetweenOperand) — `(date, nanos,
-/// offset)` — from a typed temporal array at row `i`.
-fn between_operand(
-    arr: &datafusion::arrow::array::ArrayRef,
-    i: usize,
-) -> Option<crate::temporal::BetweenOperand> {
-    use datafusion::arrow::array::{Array, StructArray, Time64NanosecondArray};
-    use datafusion::arrow::datatypes::TimeUnit;
-    if arr.is_null(i) {
-        return None;
-    }
-    match arr.data_type() {
-        DataType::Time64(TimeUnit::Nanosecond) => Some((
-            None,
-            arr.as_any()
-                .downcast_ref::<Time64NanosecondArray>()?
-                .value(i),
-            None,
-            None,
-        )),
-        DataType::Struct(_) => {
-            let s = arr.as_any().downcast_ref::<StructArray>()?;
-            if is_date_struct(arr.data_type()) {
-                Some((Some(date_struct_value(s, i)?), 0, None, None))
-            } else if is_datetime_struct(arr.data_type()) {
-                // Keep the named zone (if any) so DST is re-resolvable across a
-                // span (#1007).
-                let (days, nanos, offset, zone) = datetime_struct_parts(s, i)?;
-                Some((Some(days), nanos, Some(offset), zone))
-            } else if is_time_struct(arr.data_type()) {
-                let (nanos, offset) = time_struct_parts(s, i)?;
-                Some((None, nanos, Some(offset), None))
-            } else {
-                let (days, nanos) = localdatetime_struct_parts(s, i)?;
-                Some((Some(days), nanos, None, None))
-            }
-        }
-        _ => None,
-    }
-}
-
-impl ScalarUDFImpl for CypherDurationBetween {
-    fn name(&self) -> &'static str {
-        "cypher_duration_between"
-    }
-
-    fn signature(&self) -> &Signature {
-        &self.signature
-    }
-
-    fn return_type(&self, _arg_types: &[DataType]) -> datafusion::error::Result<DataType> {
-        Ok(DataType::Struct(
-            graphforge_ir::arrow_schema::duration_struct_fields(),
-        ))
-    }
-
-    fn invoke_with_args(
-        &self,
-        args: ScalarFunctionArgs,
-    ) -> datafusion::error::Result<ColumnarValue> {
-        use crate::temporal::{BetweenMode, duration_between};
-        use datafusion::arrow::array::{Array, StringArray};
-        use datafusion::arrow::compute::cast;
-        use datafusion::error::DataFusionError;
-
-        validate_heterogeneous_arguments(&args.args)?;
-        let rows = args.number_rows;
-        let cols = udf_argument_arrays(&args)?;
-        let mode_arr = cast(&cols[2], &DataType::Utf8).map_err(DataFusionError::from)?;
-        let modes = mode_arr.as_any().downcast_ref::<StringArray>();
-
-        let parts: Vec<Option<crate::temporal::DurationValue>> = (0..rows)
-            .map(|i| {
-                let m = modes?;
-                if m.is_null(i) {
-                    return None;
-                }
-                let mode = match m.value(i) {
-                    "duration.between" => BetweenMode::Between,
-                    "duration.inmonths" => BetweenMode::Months,
-                    "duration.indays" => BetweenMode::Days,
-                    "duration.inseconds" => BetweenMode::Seconds,
-                    _ => return None,
-                };
-                let a = between_operand(&cols[0], i)?;
-                let b = between_operand(&cols[1], i)?;
-                duration_between(&a, &b, mode)
-            })
-            .collect();
-        Ok(ColumnarValue::Array(std::sync::Arc::new(
-            build_duration_struct(&parts),
-        )))
-    }
-}
-
-// ---------------------------------------------------------------------------
-// cypher_temporal_arith UDF (temporal ± duration)
-// ---------------------------------------------------------------------------
-
-/// `temporal ± duration` (`Temporal8`): `[temporal, duration_struct, sign]` →
-/// the SAME type as the temporal operand (`return_type` echoes `arg_types[0]`).
-/// `sign` is `+1` (add) or `-1` (subtract). Dispatches on the temporal type:
-/// date adds months+days (sub-day time → whole days); localtime/time wrap the
-/// time-of-day mod 24h; localdatetime/datetime add months+days+time carrying
-/// overflow (zone offset / named zone preserved). (#920)
-static CYPHER_TEMPORAL_ARITH: LazyLock<ScalarUDF> =
-    LazyLock::new(|| ScalarUDF::new_from_impl(CypherTemporalArith::new()));
-
-#[derive(Debug, PartialEq, Eq, Hash)]
-struct CypherTemporalArith {
-    signature: Signature,
-}
-
-impl CypherTemporalArith {
-    fn new() -> Self {
-        Self {
-            signature: Signature::any(3, Volatility::Immutable),
-        }
-    }
-}
-
-impl ScalarUDFImpl for CypherTemporalArith {
-    fn name(&self) -> &'static str {
-        "cypher_temporal_arith"
-    }
-
-    fn signature(&self) -> &Signature {
-        &self.signature
-    }
-
-    fn return_type(&self, arg_types: &[DataType]) -> datafusion::error::Result<DataType> {
-        // The result is the same temporal type as the first (temporal) operand.
-        Ok(arg_types.first().cloned().unwrap_or(DataType::Null))
-    }
-
-    #[allow(
-        clippy::too_many_lines,
-        reason = "one cohesive per-temporal-type dispatch (date/localtime/time/\
-                  localdatetime/datetime) applying a signed duration"
-    )]
-    fn invoke_with_args(
-        &self,
-        args: ScalarFunctionArgs,
-    ) -> datafusion::error::Result<ColumnarValue> {
-        use crate::temporal::{
-            date_plus_duration, datetime_plus_duration, localtime_plus_duration,
-        };
-        use datafusion::arrow::array::{
-            Array, ArrayRef, Int64Array, StructArray, Time64NanosecondArray,
-        };
-        use datafusion::arrow::datatypes::TimeUnit;
-
-        validate_heterogeneous_arguments(&args.args)?;
-        let rows = args.number_rows;
-        let cols = udf_argument_arrays(&args)?;
-        let temporal = &cols[0];
-        let dur = cols[1].as_any().downcast_ref::<StructArray>();
-        let signs = cols[2].as_any().downcast_ref::<Int64Array>();
-
-        // The signed [`DurationValue`] for row `i` (None if either operand is null).
-        let signed = |i: usize| -> Option<crate::temporal::DurationValue> {
-            let (d, sg) = (dur?, signs?);
-            if d.is_null(i) || sg.is_null(i) {
-                return None;
-            }
-            let dv = duration_struct_parts(d, i)?;
-            Some(if sg.value(i) < 0 {
-                crate::temporal::DurationValue {
-                    months: -dv.months,
-                    days: -dv.days,
-                    seconds: -dv.seconds,
-                    nanos: -dv.nanos,
-                }
-            } else {
-                dv
-            })
-        };
-        // Total signed sub-day nanoseconds of a duration (time-of-day arithmetic).
-        // Sub-day nanoseconds of a duration for time-of-day (mod-24h) arithmetic.
-        // Reduce `seconds` mod a day FIRST so a huge duration can't overflow the
-        // `* 1e9` (the result is used mod a day anyway, so this is exact). (#1011)
-        let sub_day_nanos = |d: &crate::temporal::DurationValue| {
-            d.seconds.rem_euclid(86_400) * 1_000_000_000 + d.nanos
-        };
-
-        let out: ArrayRef = match temporal.data_type() {
-            DataType::Struct(_) if is_date_struct(temporal.data_type()) => {
-                let t = temporal.as_any().downcast_ref::<StructArray>();
-                let days: Vec<Option<i64>> = (0..rows)
-                    .map(|i| {
-                        let t = t?;
-                        let d = date_struct_value(t, i)?;
-                        let dv = signed(i)?;
-                        Some(date_plus_duration(d, &dv))
-                    })
-                    .collect();
-                std::sync::Arc::new(build_date_struct(&days))
-            }
-            DataType::Time64(TimeUnit::Nanosecond) => {
-                let t = temporal.as_any().downcast_ref::<Time64NanosecondArray>();
-                let a: Time64NanosecondArray = (0..rows)
-                    .map(|i| {
-                        let t = t?;
-                        if t.is_null(i) {
-                            return None;
-                        }
-                        let dv = signed(i)?;
-                        Some(localtime_plus_duration(t.value(i), sub_day_nanos(&dv)))
-                    })
-                    .collect();
-                std::sync::Arc::new(a)
-            }
-            DataType::Struct(_) if is_time_struct(temporal.data_type()) => {
-                let s = temporal.as_any().downcast_ref::<StructArray>();
-                let parts: Vec<Option<(i64, i32)>> = (0..rows)
-                    .map(|i| {
-                        let s = s?;
-                        let (nanos, offset) = time_struct_parts(s, i)?;
-                        let dv = signed(i)?;
-                        Some((localtime_plus_duration(nanos, sub_day_nanos(&dv)), offset))
-                    })
-                    .collect();
-                std::sync::Arc::new(build_time_struct(&parts))
-            }
-            DataType::Struct(_) if is_datetime_struct(temporal.data_type()) => {
-                let s = temporal.as_any().downcast_ref::<StructArray>();
-                let parts: Vec<DateTimeRow> = (0..rows)
-                    .map(|i| {
-                        let s = s?;
-                        let (days, nanos, offset, zone) = datetime_struct_parts(s, i)?;
-                        let dv = signed(i)?;
-                        let (date, no) = datetime_plus_duration(days, nanos, &dv);
-                        Some((date, no, offset, zone))
-                    })
-                    .collect();
-                std::sync::Arc::new(build_datetime_struct(&parts))
-            }
-            // localdatetime struct (date + time, no zone).
-            DataType::Struct(_) if is_localdatetime_struct(temporal.data_type()) => {
-                let s = temporal.as_any().downcast_ref::<StructArray>();
-                let parts: Vec<Option<(i64, i64)>> = (0..rows)
-                    .map(|i| {
-                        let s = s?;
-                        let (days, nanos) = localdatetime_struct_parts(s, i)?;
-                        let dv = signed(i)?;
-                        let (date, no) = datetime_plus_duration(days, nanos, &dv);
-                        Some((date, no))
-                    })
-                    .collect();
-                std::sync::Arc::new(build_localdatetime_struct(&parts))
-            }
-            other => {
-                return Err(datafusion::error::DataFusionError::Internal(format!(
-                    "cypher_temporal_arith: left operand is not a temporal value ({other:?})"
-                )));
-            }
-        };
-        Ok(ColumnarValue::Array(out))
-    }
-}
-
-// ---------------------------------------------------------------------------
-// cypher_duration_add UDF (duration ± duration)
-// ---------------------------------------------------------------------------
-
-/// Runtime `duration(<string-expr>)` (`Temporal6`): parse an ISO-8601 duration
-/// string per row into a `Struct{months, days, seconds, nanos}` (null on unparseable or
-/// null input), the inverse of the `toString` render. Used when the argument is
-/// not a constant (e.g. `duration(toString(d))`). (#920)
-static CYPHER_DURATION_PARSE: LazyLock<ScalarUDF> =
-    LazyLock::new(|| ScalarUDF::new_from_impl(CypherDurationParse::new()));
-
-#[derive(Debug, PartialEq, Eq, Hash)]
-struct CypherDurationParse {
-    signature: Signature,
-}
-
-impl CypherDurationParse {
-    fn new() -> Self {
-        Self {
-            signature: Signature::any(1, Volatility::Immutable),
-        }
-    }
-}
-
-impl ScalarUDFImpl for CypherDurationParse {
-    fn name(&self) -> &'static str {
-        "cypher_duration_parse"
-    }
-    fn signature(&self) -> &Signature {
-        &self.signature
-    }
-    fn return_type(&self, _arg_types: &[DataType]) -> datafusion::error::Result<DataType> {
-        Ok(DataType::Struct(
-            graphforge_ir::arrow_schema::duration_struct_fields(),
-        ))
-    }
-
-    fn invoke_with_args(
-        &self,
-        args: ScalarFunctionArgs,
-    ) -> datafusion::error::Result<ColumnarValue> {
-        use datafusion::arrow::array::{Array, StringArray};
-        use datafusion::arrow::compute::cast;
-
-        validate_heterogeneous_arguments(&args.args)?;
-        let rows = args.number_rows;
-        let arr = cast(&args.args[0].to_array(rows)?, &DataType::Utf8)?;
-        let s = arr.as_any().downcast_ref::<StringArray>();
-        let parts: Vec<Option<crate::temporal::DurationValue>> = (0..rows)
-            .map(|i| {
-                let s = s?;
-                if s.is_null(i) {
-                    return Option::None;
-                }
-                crate::temporal::duration_value_from_str(s.value(i))
-            })
-            .collect();
-        Ok(ColumnarValue::Array(std::sync::Arc::new(
-            build_duration_struct(&parts),
-        )))
-    }
-}
-
-/// `duration ± duration` (`Temporal8`): `[a, b, sign]` → component-wise
-/// `(a.months + sign·b.months, …days, …nanos)` as a duration struct. (#920)
-static CYPHER_DURATION_ADD: LazyLock<ScalarUDF> =
-    LazyLock::new(|| ScalarUDF::new_from_impl(CypherDurationAdd::new()));
-
-#[derive(Debug, PartialEq, Eq, Hash)]
-struct CypherDurationAdd {
-    signature: Signature,
-}
-
-impl CypherDurationAdd {
-    fn new() -> Self {
-        Self {
-            signature: Signature::any(3, Volatility::Immutable),
-        }
-    }
-}
-
-impl ScalarUDFImpl for CypherDurationAdd {
-    fn name(&self) -> &'static str {
-        "cypher_duration_add"
-    }
-
-    fn signature(&self) -> &Signature {
-        &self.signature
-    }
-
-    fn return_type(&self, _arg_types: &[DataType]) -> datafusion::error::Result<DataType> {
-        Ok(DataType::Struct(
-            graphforge_ir::arrow_schema::duration_struct_fields(),
-        ))
-    }
-
-    fn invoke_with_args(
-        &self,
-        args: ScalarFunctionArgs,
-    ) -> datafusion::error::Result<ColumnarValue> {
-        use datafusion::arrow::array::{Array, Int64Array, StructArray};
-
-        validate_heterogeneous_arguments(&args.args)?;
-        let rows = args.number_rows;
-        let cols = udf_argument_arrays(&args)?;
-        let a = cols[0].as_any().downcast_ref::<StructArray>();
-        let b = cols[1].as_any().downcast_ref::<StructArray>();
-        let signs = cols[2].as_any().downcast_ref::<Int64Array>();
-
-        let parts: Vec<Option<crate::temporal::DurationValue>> = (0..rows)
-            .map(|i| {
-                let (a, b, sg) = (a?, b?, signs?);
-                let av = duration_struct_parts(a, i)?;
-                let bv = duration_struct_parts(b, i)?;
-                let s: i64 = if sg.is_null(i) || sg.value(i) >= 0 {
-                    1
-                } else {
-                    -1
-                };
-                // Add componentwise and normalise the nanos carry into seconds —
-                // WITHOUT forming a `seconds * 1e9` total (which would overflow
-                // i64 for combined sub-day spans > ~292 years, defeating the
-                // widened `seconds` field). `nanos` sums into (-1e9, 2e9), so
-                // div/rem_euclid re-canonicalise to a non-negative `[0, 1e9)`. (#1011)
-                let nanos_sum = av.nanos + s * bv.nanos;
-                let seconds = av.seconds + s * bv.seconds + nanos_sum.div_euclid(1_000_000_000);
-                Some(crate::temporal::DurationValue {
-                    months: av.months + s * bv.months,
-                    days: av.days + s * bv.days,
-                    seconds,
-                    nanos: nanos_sum.rem_euclid(1_000_000_000),
-                })
-            })
-            .collect();
-        Ok(ColumnarValue::Array(std::sync::Arc::new(
-            build_duration_struct(&parts),
-        )))
-    }
-}
-
-/// `duration * number` / `duration / number` (#920 Temporal8 [7]). Args are
-/// `[duration_struct, number, is_div]`; scales each component and re-normalises
-/// via [`crate::temporal::scale_duration`] (fractional months → days → time).
-static CYPHER_DURATION_SCALE: LazyLock<ScalarUDF> =
-    LazyLock::new(|| ScalarUDF::new_from_impl(CypherDurationScale::new()));
-
-#[derive(Debug, PartialEq, Eq, Hash)]
-struct CypherDurationScale {
-    signature: Signature,
-}
-
-impl CypherDurationScale {
-    fn new() -> Self {
-        Self {
-            signature: Signature::any(3, Volatility::Immutable),
-        }
-    }
-}
-
-impl ScalarUDFImpl for CypherDurationScale {
-    fn name(&self) -> &'static str {
-        "cypher_duration_scale"
-    }
-    fn signature(&self) -> &Signature {
-        &self.signature
-    }
-    fn return_type(&self, _arg_types: &[DataType]) -> datafusion::error::Result<DataType> {
-        Ok(DataType::Struct(
-            graphforge_ir::arrow_schema::duration_struct_fields(),
-        ))
-    }
-
-    fn invoke_with_args(
-        &self,
-        args: ScalarFunctionArgs,
-    ) -> datafusion::error::Result<ColumnarValue> {
-        use datafusion::arrow::array::{Array, BooleanArray, Float64Array, StructArray};
-        use datafusion::arrow::compute::cast;
-
-        validate_heterogeneous_arguments(&args.args)?;
-        let rows = args.number_rows;
-        let cols = udf_argument_arrays(&args)?;
-        let dur = cols[0].as_any().downcast_ref::<StructArray>();
-        // The numeric factor may arrive as any int/float width — cast to f64.
-        let num = cast(&cols[1], &DataType::Float64)?;
-        let num = num.as_any().downcast_ref::<Float64Array>();
-        let is_div = cols[2].as_any().downcast_ref::<BooleanArray>();
-
-        let parts: Vec<Option<crate::temporal::DurationValue>> = (0..rows)
-            .map(|i| {
-                let (dur, num, is_div) = (dur?, num?, is_div?);
-                if num.is_null(i) {
-                    return Option::None; // duration ∘ null = null
-                }
-                let dv = duration_struct_parts(dur, i)?;
-                let divide = !is_div.is_null(i) && is_div.value(i);
-                Some(crate::temporal::scale_duration(&dv, num.value(i), divide))
-            })
-            .collect();
-        Ok(ColumnarValue::Array(std::sync::Arc::new(
-            build_duration_struct(&parts),
-        )))
-    }
-}
-
 // ---------------------------------------------------------------------------
 // cypher_quantifier UDF (all/any/none/single)
 // ---------------------------------------------------------------------------
@@ -8418,1710 +7621,6 @@ impl ScalarUDFImpl for CypherListComp {
 }
 
 // ---------------------------------------------------------------------------
-// cypher_date_project UDF
-// ---------------------------------------------------------------------------
-
-fn udf_argument_arrays(
-    args: &ScalarFunctionArgs,
-) -> datafusion::error::Result<Vec<datafusion::arrow::array::ArrayRef>> {
-    args.args
-        .iter()
-        .map(|value| value.to_array(args.number_rows))
-        .collect()
-}
-
-fn cast_argument_arrays(
-    arrays: &[datafusion::arrow::array::ArrayRef],
-    data_type: &DataType,
-) -> datafusion::error::Result<Vec<datafusion::arrow::array::ArrayRef>> {
-    arrays
-        .iter()
-        .map(|array| datafusion::arrow::compute::cast(array, data_type).map_err(Into::into))
-        .collect()
-}
-
-/// `date`-from-value projection (`Temporal3`): `date(base)` / `date({date: base,
-/// …overrides})`. Args are `[base, year, month, day, week, dayOfWeek,
-/// ordinalDay, quarter, dayOfQuarter]` — `base` is a `Date32` or an ISO date/
-/// datetime string, the eight overrides are nullable integers (null ⇒ keep the
-/// base's component). Returns `Date32`. (ADR 0009 / #920)
-static CYPHER_DATE_PROJECT: LazyLock<ScalarUDF> =
-    LazyLock::new(|| ScalarUDF::new_from_impl(CypherDateProject::new()));
-
-#[derive(Debug, PartialEq, Eq, Hash)]
-struct CypherDateProject {
-    signature: Signature,
-}
-
-impl CypherDateProject {
-    fn new() -> Self {
-        Self {
-            signature: Signature::any(9, Volatility::Immutable),
-        }
-    }
-}
-
-impl ScalarUDFImpl for CypherDateProject {
-    fn name(&self) -> &'static str {
-        "cypher_date_project"
-    }
-
-    fn signature(&self) -> &Signature {
-        &self.signature
-    }
-
-    fn return_type(&self, _arg_types: &[DataType]) -> datafusion::error::Result<DataType> {
-        Ok(DataType::Struct(
-            graphforge_ir::arrow_schema::date_struct_fields(),
-        ))
-    }
-
-    fn invoke_with_args(
-        &self,
-        args: ScalarFunctionArgs,
-    ) -> datafusion::error::Result<ColumnarValue> {
-        use crate::temporal::{DateOverrides, parse_date_or_datetime_prefix};
-        use datafusion::arrow::array::{Array, StringArray, StructArray};
-        use datafusion::arrow::compute::cast;
-
-        validate_heterogeneous_arguments(&args.args)?;
-        let rows = args.number_rows;
-        let cols = udf_argument_arrays(&args)?;
-        // A typed temporal-struct base (date / localdatetime / time / datetime) is
-        // read directly; a string base is cast to `Utf8` (handling `Utf8View`).
-        let base = if is_date_struct(cols[0].data_type())
-            || is_localdatetime_struct(cols[0].data_type())
-            || is_time_struct(cols[0].data_type())
-            || is_datetime_struct(cols[0].data_type())
-        {
-            std::sync::Arc::clone(&cols[0])
-        } else {
-            cast(&cols[0], &DataType::Utf8).map_err(datafusion::error::DataFusionError::from)?
-        };
-        // Overrides: cast each to Int64 once (a null/absent override stays null).
-        let ov: Vec<_> = cols[1..9]
-            .iter()
-            .map(|c| cast(c, &DataType::Int64))
-            .collect::<std::result::Result<Vec<_>, _>>()
-            .map_err(datafusion::error::DataFusionError::from)?;
-
-        let base_date = |i: usize| -> Option<i64> {
-            if base.is_null(i) {
-                return None;
-            }
-            match base.data_type() {
-                DataType::Struct(_) => {
-                    let s = base.as_any().downcast_ref::<StructArray>()?;
-                    if is_date_struct(base.data_type()) {
-                        date_struct_value(s, i)
-                    } else {
-                        // A `localdatetime`/`datetime` value — take its date component.
-                        Some(localdatetime_struct_parts(s, i)?.0)
-                    }
-                }
-                DataType::Utf8 => parse_date_or_datetime_prefix(
-                    base.as_any().downcast_ref::<StringArray>()?.value(i),
-                ),
-                _ => None,
-            }
-        };
-
-        let out: Vec<Option<i64>> = (0..rows)
-            .map(|i| {
-                let overrides = DateOverrides {
-                    year: optional_i64_at(&ov[0], i),
-                    month: optional_i64_at(&ov[1], i),
-                    day: optional_i64_at(&ov[2], i),
-                    week: optional_i64_at(&ov[3], i),
-                    day_of_week: optional_i64_at(&ov[4], i),
-                    ordinal_day: optional_i64_at(&ov[5], i),
-                    quarter: optional_i64_at(&ov[6], i),
-                    day_of_quarter: optional_i64_at(&ov[7], i),
-                };
-                crate::temporal::project_date(base_date(i)?, &overrides)
-            })
-            .collect();
-        Ok(ColumnarValue::Array(std::sync::Arc::new(
-            build_date_struct(&out),
-        )))
-    }
-}
-
-// ---------------------------------------------------------------------------
-// cypher_localtime_project UDF
-// ---------------------------------------------------------------------------
-
-/// `localtime`-from-value projection (`Temporal3`): `localtime(base)` /
-/// `localtime({time: base, …overrides})`. Args are `[base, hour, minute, second,
-/// millisecond, microsecond, nanosecond]` — `base` is a `Time64(Nanosecond)` or
-/// any ISO temporal string (its time-of-day is extracted), the six overrides are
-/// nullable integers (null ⇒ keep the base's component). Returns
-/// `Time64(Nanosecond)`. (ADR 0009)
-static CYPHER_LOCALTIME_PROJECT: LazyLock<ScalarUDF> =
-    LazyLock::new(|| ScalarUDF::new_from_impl(CypherLocalTimeProject::new()));
-
-#[derive(Debug, PartialEq, Eq, Hash)]
-struct CypherLocalTimeProject {
-    signature: Signature,
-}
-
-impl CypherLocalTimeProject {
-    fn new() -> Self {
-        Self {
-            signature: Signature::any(7, Volatility::Immutable),
-        }
-    }
-}
-
-impl ScalarUDFImpl for CypherLocalTimeProject {
-    fn name(&self) -> &'static str {
-        "cypher_localtime_project"
-    }
-
-    fn signature(&self) -> &Signature {
-        &self.signature
-    }
-
-    fn return_type(&self, _arg_types: &[DataType]) -> datafusion::error::Result<DataType> {
-        Ok(DataType::Time64(
-            datafusion::arrow::datatypes::TimeUnit::Nanosecond,
-        ))
-    }
-
-    fn invoke_with_args(
-        &self,
-        args: ScalarFunctionArgs,
-    ) -> datafusion::error::Result<ColumnarValue> {
-        use crate::temporal::{LocalTimeOverrides, project_localtime, time_of_day_nanos_any};
-        use datafusion::arrow::array::{
-            Array, ArrayRef, StringArray, StructArray, Time64NanosecondArray,
-        };
-        use datafusion::arrow::compute::cast;
-        use datafusion::arrow::datatypes::TimeUnit;
-        use datafusion::error::DataFusionError;
-
-        validate_heterogeneous_arguments(&args.args)?;
-        let rows = args.number_rows;
-        let cols = udf_argument_arrays(&args)?;
-        // A `Time64(Nanosecond)` or `localdatetime`-struct base is read directly;
-        // any other (string) base is cast to `Utf8` (handling `Utf8View`).
-        let base: ArrayRef =
-            if matches!(cols[0].data_type(), DataType::Time64(TimeUnit::Nanosecond))
-                || is_localdatetime_struct(cols[0].data_type())
-                || is_time_struct(cols[0].data_type())
-                || is_datetime_struct(cols[0].data_type())
-            {
-                std::sync::Arc::clone(&cols[0])
-            } else {
-                cast(&cols[0], &DataType::Utf8).map_err(DataFusionError::from)?
-            };
-        let ov = cast_argument_arrays(&cols[1..7], &DataType::Int64)?;
-
-        let base_nanos = |i: usize| -> Option<i64> {
-            if base.is_null(i) {
-                return None;
-            }
-            match base.data_type() {
-                DataType::Time64(TimeUnit::Nanosecond) => Some(
-                    base.as_any()
-                        .downcast_ref::<Time64NanosecondArray>()?
-                        .value(i),
-                ),
-                // A `localdatetime` or `time` value — take its time-of-day.
-                DataType::Struct(_) => {
-                    let s = base.as_any().downcast_ref::<StructArray>()?;
-                    if is_time_struct(base.data_type()) {
-                        Some(time_struct_parts(s, i)?.0)
-                    } else {
-                        Some(localdatetime_struct_parts(s, i)?.1)
-                    }
-                }
-                DataType::Utf8 => {
-                    time_of_day_nanos_any(base.as_any().downcast_ref::<StringArray>()?.value(i))
-                }
-                _ => None,
-            }
-        };
-
-        let out: Time64NanosecondArray = (0..rows)
-            .map(|i| {
-                let overrides = LocalTimeOverrides {
-                    hour: optional_i64_at(&ov[0], i),
-                    minute: optional_i64_at(&ov[1], i),
-                    second: optional_i64_at(&ov[2], i),
-                    millisecond: optional_i64_at(&ov[3], i),
-                    microsecond: optional_i64_at(&ov[4], i),
-                    nanosecond: optional_i64_at(&ov[5], i),
-                };
-                project_localtime(base_nanos(i)?, &overrides)
-            })
-            .collect();
-        Ok(ColumnarValue::Array(std::sync::Arc::new(out)))
-    }
-}
-
-// ---------------------------------------------------------------------------
-// cypher_localtime_truncate UDF
-// ---------------------------------------------------------------------------
-
-/// `localtime.truncate(unit, value, map)` (`Temporal9`): truncate `value`'s
-/// time-of-day to `unit`, then apply the override `map`. Args are `[value, unit,
-/// hour, minute, second, millisecond, microsecond, nanosecond]` — `value` is a
-/// `Time64(ns)` / `localdatetime`/`time`/`datetime` struct / ISO string, `unit` a
-/// string, the six overrides nullable integers. Returns `Time64(ns)`. (#920)
-static CYPHER_LOCALTIME_TRUNCATE: LazyLock<ScalarUDF> =
-    LazyLock::new(|| ScalarUDF::new_from_impl(CypherLocalTimeTruncate::new()));
-
-#[derive(Debug, PartialEq, Eq, Hash)]
-struct CypherLocalTimeTruncate {
-    signature: Signature,
-}
-
-impl CypherLocalTimeTruncate {
-    fn new() -> Self {
-        Self {
-            signature: Signature::any(8, Volatility::Immutable),
-        }
-    }
-}
-
-impl ScalarUDFImpl for CypherLocalTimeTruncate {
-    fn name(&self) -> &'static str {
-        "cypher_localtime_truncate"
-    }
-
-    fn signature(&self) -> &Signature {
-        &self.signature
-    }
-
-    fn return_type(&self, _arg_types: &[DataType]) -> datafusion::error::Result<DataType> {
-        Ok(DataType::Time64(
-            datafusion::arrow::datatypes::TimeUnit::Nanosecond,
-        ))
-    }
-
-    fn invoke_with_args(
-        &self,
-        args: ScalarFunctionArgs,
-    ) -> datafusion::error::Result<ColumnarValue> {
-        use crate::temporal::{
-            LocalTimeOverrides, project_localtime, time_of_day_nanos_any, truncate_time_nanos,
-        };
-        use datafusion::arrow::array::{
-            Array, ArrayRef, StringArray, StructArray, Time64NanosecondArray,
-        };
-        use datafusion::arrow::compute::cast;
-        use datafusion::arrow::datatypes::TimeUnit;
-        use datafusion::error::DataFusionError;
-
-        validate_heterogeneous_arguments(&args.args)?;
-        let rows = args.number_rows;
-        let cols = udf_argument_arrays(&args)?;
-        let base: ArrayRef =
-            if matches!(cols[0].data_type(), DataType::Time64(TimeUnit::Nanosecond))
-                || is_localdatetime_struct(cols[0].data_type())
-                || is_time_struct(cols[0].data_type())
-                || is_datetime_struct(cols[0].data_type())
-            {
-                std::sync::Arc::clone(&cols[0])
-            } else {
-                cast(&cols[0], &DataType::Utf8).map_err(DataFusionError::from)?
-            };
-        let units_arr = cast(&cols[1], &DataType::Utf8).map_err(DataFusionError::from)?;
-        let units = units_arr.as_any().downcast_ref::<StringArray>();
-        let ov = cast_argument_arrays(&cols[2..8], &DataType::Int64)?;
-
-        let base_nanos = |i: usize| -> Option<i64> {
-            if base.is_null(i) {
-                return None;
-            }
-            match base.data_type() {
-                DataType::Time64(TimeUnit::Nanosecond) => Some(
-                    base.as_any()
-                        .downcast_ref::<Time64NanosecondArray>()?
-                        .value(i),
-                ),
-                DataType::Struct(_) => {
-                    let s = base.as_any().downcast_ref::<StructArray>()?;
-                    if is_time_struct(base.data_type()) {
-                        Some(time_struct_parts(s, i)?.0)
-                    } else {
-                        Some(localdatetime_struct_parts(s, i)?.1)
-                    }
-                }
-                DataType::Utf8 => {
-                    time_of_day_nanos_any(base.as_any().downcast_ref::<StringArray>()?.value(i))
-                }
-                _ => None,
-            }
-        };
-
-        let out: Time64NanosecondArray = (0..rows)
-            .map(|i| {
-                let u = units?;
-                if u.is_null(i) {
-                    return None;
-                }
-                let truncated = truncate_time_nanos(base_nanos(i)?, u.value(i))?;
-                let overrides = LocalTimeOverrides {
-                    hour: optional_i64_at(&ov[0], i),
-                    minute: optional_i64_at(&ov[1], i),
-                    second: optional_i64_at(&ov[2], i),
-                    millisecond: optional_i64_at(&ov[3], i),
-                    microsecond: optional_i64_at(&ov[4], i),
-                    nanosecond: optional_i64_at(&ov[5], i),
-                };
-                project_localtime(truncated, &overrides)
-            })
-            .collect();
-        Ok(ColumnarValue::Array(std::sync::Arc::new(out)))
-    }
-}
-
-// ---------------------------------------------------------------------------
-// cypher_localdatetime_project UDF
-// ---------------------------------------------------------------------------
-
-/// The Arrow fields of a standalone `date` value — `Struct{epoch_day: Int64}`
-/// (ADR 0012). A one-field struct (not a bare `Int64`) so a `date` is
-/// self-describing on storage decode — a plain integer property would be
-/// indistinguishable — while spanning the full openCypher year range (#1011).
-fn date_fields() -> datafusion::arrow::datatypes::Fields {
-    graphforge_ir::arrow_schema::date_struct_fields()
-}
-
-/// True if `dt` is the standalone `date` struct type (`Struct{epoch_day: Int64}`),
-/// distinguished from the other temporal structs by its single field name.
-fn is_date_struct(dt: &DataType) -> bool {
-    matches!(dt, DataType::Struct(fields)
-        if fields.len() == 1
-            && fields[0].name() == "epoch_day"
-            && *fields[0].data_type() == DataType::Int64)
-}
-
-/// Build a standalone `date` struct array from per-row i64 epoch-days (`None` ⇒ a
-/// null row).
-fn build_date_struct(rows: &[Option<i64>]) -> datafusion::arrow::array::StructArray {
-    use datafusion::arrow::array::Int64Array;
-    use datafusion::arrow::buffer::NullBuffer;
-    let days: Int64Array = rows.iter().copied().collect();
-    let nulls = rows.iter().map(Option::is_some).collect::<NullBuffer>();
-    datafusion::arrow::array::StructArray::new(
-        date_fields(),
-        vec![std::sync::Arc::new(days)],
-        Some(nulls),
-    )
-}
-
-/// A standalone `date` scalar (`None` ⇒ a null value).
-fn date_scalar(days: Option<i64>) -> ScalarValue {
-    ScalarValue::Struct(std::sync::Arc::new(build_date_struct(&[days])))
-}
-
-/// The i64 epoch-day of a standalone `date` struct row (`None` for a null row).
-fn date_struct_value(arr: &datafusion::arrow::array::StructArray, i: usize) -> Option<i64> {
-    use datafusion::arrow::array::{Array, Int64Array};
-    if arr.is_null(i) {
-        return None;
-    }
-    let days = arr.column(0).as_any().downcast_ref::<Int64Array>()?;
-    days.is_valid(i).then(|| days.value(i))
-}
-
-/// Read one nullable Int64 override from an already-normalized Arrow column.
-/// Temporal projectors cast override columns once before their row loop, so a
-/// failed physical downcast or a null row has the same absent-override meaning.
-fn optional_i64_at(array: &datafusion::arrow::array::ArrayRef, row: usize) -> Option<i64> {
-    use datafusion::arrow::array::{Array, Int64Array};
-
-    let values = array.as_any().downcast_ref::<Int64Array>()?;
-    (!values.is_null(row)).then(|| values.value(row))
-}
-
-/// The Arrow fields of a `localdatetime` value — `Struct{date: Int64, time:
-/// Time64(Nanosecond)}`. `date` is first so DataFusion's row-format sort orders
-/// chronologically (date, then time-of-day); `date` is i64 days (#1011).
-fn localdatetime_fields() -> datafusion::arrow::datatypes::Fields {
-    graphforge_ir::arrow_schema::localdatetime_struct_fields()
-}
-
-/// True if `dt` is the `localdatetime` struct type (used to dispatch base
-/// extraction and rendering without colliding with user maps, whose `date`/
-/// `time` fields would not have these exact `Int64`/`Time64` types).
-fn is_localdatetime_struct(dt: &DataType) -> bool {
-    use datafusion::arrow::datatypes::TimeUnit;
-    matches!(dt, DataType::Struct(fields)
-        if fields.len() == 2
-            && fields[0].name() == "date"
-            && *fields[0].data_type() == DataType::Int64
-            && fields[1].name() == "time"
-            && *fields[1].data_type() == DataType::Time64(TimeUnit::Nanosecond))
-}
-
-/// Build a `localdatetime` struct array from per-row `(date_days, nanos_of_day)`
-/// (`None` ⇒ a null row).
-fn build_localdatetime_struct(
-    rows: &[Option<(i64, i64)>],
-) -> datafusion::arrow::array::StructArray {
-    use datafusion::arrow::array::{Int64Array, Time64NanosecondArray};
-    use datafusion::arrow::buffer::NullBuffer;
-    let days: Int64Array = rows.iter().map(|r| r.map(|(d, _)| d)).collect();
-    let nanos: Time64NanosecondArray = rows.iter().map(|r| r.map(|(_, n)| n)).collect();
-    let nulls = rows.iter().map(Option::is_some).collect::<NullBuffer>();
-    datafusion::arrow::array::StructArray::new(
-        localdatetime_fields(),
-        vec![std::sync::Arc::new(days), std::sync::Arc::new(nanos)],
-        Some(nulls),
-    )
-}
-
-/// A `localdatetime` scalar (`None` ⇒ a null value).
-fn localdatetime_scalar(parts: Option<(i64, i64)>) -> ScalarValue {
-    ScalarValue::Struct(std::sync::Arc::new(build_localdatetime_struct(&[parts])))
-}
-
-/// True if `dt` is the typed `duration` struct (`Struct{months, days, seconds, nanos}`,
-/// ADR 0009) — distinguished from the other temporal structs by its field names.
-fn is_duration_struct(dt: &DataType) -> bool {
-    matches!(dt, DataType::Struct(fields)
-        if fields.len() == 4
-            && fields[0].name() == "months"
-            && fields[1].name() == "days"
-            && fields[2].name() == "seconds"
-            && fields[3].name() == "nanos")
-}
-
-/// Whether a function name is a temporal clock accessor —
-/// `<type>.transaction` / `.statement` / `.realtime` for an instant type
-/// (`date`/`localtime`/`time`/`localdatetime`/`datetime`). (#920)
-fn is_temporal_clock_fn(name: &str) -> bool {
-    // Cypher function names are case-insensitive (`Date.Realtime` ≡ `date.realtime`).
-    matches!(
-        name.to_ascii_lowercase().split_once('.'),
-        Some((
-            "date" | "localtime" | "time" | "localdatetime" | "datetime",
-            "transaction" | "statement" | "realtime",
-        ))
-    )
-}
-
-/// The typed null for a temporal constructor / clock function — `date` →
-/// `Date32(None)`, `localtime` → `Time64(None)`, etc. — so null propagation
-/// preserves the Arrow temporal type rather than erasing it to `Null` (which
-/// would defeat downstream `is_temporal_typed` checks). The base name is the
-/// part before any `.clock` suffix, matched case-insensitively. (#920)
-fn temporal_null_scalar(name: &str) -> ScalarValue {
-    let lower = name.to_ascii_lowercase();
-    let base = lower.split('.').next().unwrap_or(&lower);
-    match base {
-        "date" => date_scalar(None),
-        "localtime" => ScalarValue::Time64Nanosecond(None),
-        "time" => time_scalar(None),
-        "localdatetime" => localdatetime_scalar(None),
-        "datetime" => datetime_scalar(None),
-        "duration" => duration_scalar(None),
-        _ => ScalarValue::Null,
-    }
-}
-
-/// Build a `duration` struct array from per-row [`DurationValue`]s (`None` ⇒ a
-/// null row). The on-disk + query representation of a Cypher duration —
-/// `Struct{months,days,seconds,nanos}` all Int64 (Parquet cannot persist Arrow
-/// `Interval(MonthDayNano)`). (#920/#1011)
-fn build_duration_struct(
-    rows: &[Option<crate::temporal::DurationValue>],
-) -> datafusion::arrow::array::StructArray {
-    use datafusion::arrow::array::Int64Array;
-    use datafusion::arrow::buffer::NullBuffer;
-    let months: Int64Array = rows.iter().map(|r| r.map(|d| d.months)).collect();
-    let days: Int64Array = rows.iter().map(|r| r.map(|d| d.days)).collect();
-    let seconds: Int64Array = rows.iter().map(|r| r.map(|d| d.seconds)).collect();
-    let nanos: Int64Array = rows.iter().map(|r| r.map(|d| d.nanos)).collect();
-    let nulls = rows.iter().map(Option::is_some).collect::<NullBuffer>();
-    datafusion::arrow::array::StructArray::new(
-        graphforge_ir::arrow_schema::duration_struct_fields(),
-        vec![
-            std::sync::Arc::new(months),
-            std::sync::Arc::new(days),
-            std::sync::Arc::new(seconds),
-            std::sync::Arc::new(nanos),
-        ],
-        Some(nulls),
-    )
-}
-
-/// Extract a [`DurationValue`] from a `duration` struct array at row `i`
-/// (`None` for a null row). (#920/#1011)
-fn duration_struct_parts(
-    arr: &datafusion::arrow::array::StructArray,
-    i: usize,
-) -> Option<crate::temporal::DurationValue> {
-    use datafusion::arrow::array::{Array, Int64Array};
-    if arr.is_null(i) {
-        return None;
-    }
-    let col = |idx: usize| arr.column(idx).as_any().downcast_ref::<Int64Array>();
-    Some(crate::temporal::DurationValue {
-        months: col(0)?.value(i),
-        days: col(1)?.value(i),
-        seconds: col(2)?.value(i),
-        nanos: col(3)?.value(i),
-    })
-}
-
-/// Build a typed `duration` scalar from a [`DurationValue`] (`None` ⇒ null). (#920)
-fn duration_scalar(parts: Option<crate::temporal::DurationValue>) -> ScalarValue {
-    ScalarValue::Struct(std::sync::Arc::new(build_duration_struct(&[parts])))
-}
-
-/// A sub-day-only [`DurationValue`] from whole `seconds` + non-negative
-/// `nanos`-of-second (no month/day part) — for the native-Arrow duration arms. (#1011)
-fn dur_secs_nanos(seconds: i64, nanos: i64) -> crate::temporal::DurationValue {
-    crate::temporal::DurationValue {
-        months: 0,
-        days: 0,
-        seconds,
-        nanos,
-    }
-}
-
-/// Convert a [`DurationValue`] to an `IrLiteral::Duration` (storage form). (#1011)
-fn duration_value_to_ir(d: crate::temporal::DurationValue) -> IrLiteral {
-    IrLiteral::Duration {
-        months: d.months,
-        days: d.days,
-        seconds: d.seconds,
-        nanos: d.nanos,
-    }
-}
-
-/// Extract `(date_days, nanos_of_day)` from a `localdatetime` struct array at
-/// row `i` (`None` for a null row). Also used to read the local date+time of a
-/// `datetime` struct (whose leading two fields are the same `Int64`+`Time64`),
-/// dropping its zone — the correct semantics for `date`/`localtime`/
-/// `localdatetime` projections from a `datetime`.
-fn localdatetime_struct_parts(
-    arr: &datafusion::arrow::array::StructArray,
-    i: usize,
-) -> Option<(i64, i64)> {
-    use datafusion::arrow::array::{Array, Int64Array, Time64NanosecondArray};
-    if arr.is_null(i) {
-        return None;
-    }
-    let d = arr.column(0).as_any().downcast_ref::<Int64Array>()?;
-    let t = arr
-        .column(1)
-        .as_any()
-        .downcast_ref::<Time64NanosecondArray>()?;
-    (!d.is_null(i) && !t.is_null(i)).then(|| (d.value(i), t.value(i)))
-}
-
-/// `localdatetime`-from-value projection (`Temporal3`). Args are `[date_source,
-/// time_source, year, month, day, week, dayOfWeek, ordinalDay, quarter,
-/// dayOfQuarter, hour, minute, second, millisecond, microsecond, nanosecond]`.
-/// `date_source`/`time_source` are the lowered `datetime`/`date`/`time` anchors
-/// (a `Date32`/`Time64`/`localdatetime`-struct/temporal-string, or null),
-/// interpreted as a date and a time-of-day respectively (a null date ⇒ epoch, a
-/// null time ⇒ midnight). The 14 overrides are nullable integers (null ⇒ keep
-/// the base's component). Returns the `localdatetime` struct. (ADR 0009)
-static CYPHER_LOCALDATETIME_PROJECT: LazyLock<ScalarUDF> =
-    LazyLock::new(|| ScalarUDF::new_from_impl(CypherLocalDateTimeProject::new()));
-
-#[derive(Debug, PartialEq, Eq, Hash)]
-struct CypherLocalDateTimeProject {
-    signature: Signature,
-}
-
-impl CypherLocalDateTimeProject {
-    fn new() -> Self {
-        Self {
-            signature: Signature::any(16, Volatility::Immutable),
-        }
-    }
-}
-
-impl ScalarUDFImpl for CypherLocalDateTimeProject {
-    fn name(&self) -> &'static str {
-        "cypher_localdatetime_project"
-    }
-
-    fn signature(&self) -> &Signature {
-        &self.signature
-    }
-
-    fn return_type(&self, _arg_types: &[DataType]) -> datafusion::error::Result<DataType> {
-        Ok(DataType::Struct(localdatetime_fields()))
-    }
-
-    #[allow(
-        clippy::too_many_lines,
-        reason = "one cohesive per-row projection: two typed base extractions \
-                  (date + time) plus 14 component overrides — splitting it would \
-                  scatter the row logic across helpers without aiding clarity"
-    )]
-    fn invoke_with_args(
-        &self,
-        args: ScalarFunctionArgs,
-    ) -> datafusion::error::Result<ColumnarValue> {
-        use crate::temporal::{
-            DateOverrides, LocalTimeOverrides, parse_date_or_datetime_prefix, project_date,
-            project_localtime, time_of_day_nanos_any,
-        };
-        use datafusion::arrow::array::{
-            Array, ArrayRef, StringArray, StructArray, Time64NanosecondArray,
-        };
-        use datafusion::arrow::compute::cast;
-        use datafusion::arrow::datatypes::TimeUnit;
-        use datafusion::error::DataFusionError;
-
-        validate_heterogeneous_arguments(&args.args)?;
-        let rows = args.number_rows;
-        let cols = udf_argument_arrays(&args)?;
-        // Typed sources (`date`/`localdatetime`/`time`/`datetime` struct or
-        // `Time64`) are read directly; a string source is cast to `Utf8`
-        // (handling `Utf8View`).
-        let typed_or_utf8 = |a: &ArrayRef| -> datafusion::error::Result<ArrayRef> {
-            if matches!(a.data_type(), DataType::Time64(TimeUnit::Nanosecond))
-                || is_date_struct(a.data_type())
-                || is_localdatetime_struct(a.data_type())
-                || is_time_struct(a.data_type())
-                || is_datetime_struct(a.data_type())
-            {
-                Ok(std::sync::Arc::clone(a))
-            } else {
-                cast(a, &DataType::Utf8).map_err(DataFusionError::from)
-            }
-        };
-        let date_src = typed_or_utf8(&cols[0])?;
-        let time_src = typed_or_utf8(&cols[1])?;
-        let ov = cast_argument_arrays(&cols[2..16], &DataType::Int64)?;
-
-        let base_date = |i: usize| -> Option<i64> {
-            if date_src.is_null(i) {
-                return Some(0); // a missing date defaults to the epoch (day 0)
-            }
-            match date_src.data_type() {
-                DataType::Struct(_) => {
-                    let s = date_src.as_any().downcast_ref::<StructArray>()?;
-                    if is_date_struct(date_src.data_type()) {
-                        date_struct_value(s, i)
-                    } else {
-                        // A `localdatetime`/`datetime` value — take its date.
-                        Some(localdatetime_struct_parts(s, i)?.0)
-                    }
-                }
-                DataType::Utf8 => parse_date_or_datetime_prefix(
-                    date_src.as_any().downcast_ref::<StringArray>()?.value(i),
-                ),
-                _ => None,
-            }
-        };
-        let base_time = |i: usize| -> Option<i64> {
-            if time_src.is_null(i) {
-                return Some(0); // a missing time defaults to midnight
-            }
-            match time_src.data_type() {
-                DataType::Time64(TimeUnit::Nanosecond) => Some(
-                    time_src
-                        .as_any()
-                        .downcast_ref::<Time64NanosecondArray>()?
-                        .value(i),
-                ),
-                DataType::Struct(_) => {
-                    let s = time_src.as_any().downcast_ref::<StructArray>()?;
-                    // A date-only source (bare `localdatetime(date(…))`, where the
-                    // same value feeds both slots) has no time-of-day → midnight,
-                    // matching `localdatetime({date: d})`. (A time-only source in
-                    // the *date* slot correctly stays null — no date to fabricate.)
-                    if is_date_struct(time_src.data_type()) {
-                        Some(0)
-                    } else if is_time_struct(time_src.data_type()) {
-                        Some(time_struct_parts(s, i)?.0)
-                    } else {
-                        Some(localdatetime_struct_parts(s, i)?.1)
-                    }
-                }
-                DataType::Utf8 => {
-                    time_of_day_nanos_any(time_src.as_any().downcast_ref::<StringArray>()?.value(i))
-                }
-                _ => None,
-            }
-        };
-
-        let parts: Vec<Option<(i64, i64)>> = (0..rows)
-            .map(|i| {
-                let date_overrides = DateOverrides {
-                    year: optional_i64_at(&ov[0], i),
-                    month: optional_i64_at(&ov[1], i),
-                    day: optional_i64_at(&ov[2], i),
-                    week: optional_i64_at(&ov[3], i),
-                    day_of_week: optional_i64_at(&ov[4], i),
-                    ordinal_day: optional_i64_at(&ov[5], i),
-                    quarter: optional_i64_at(&ov[6], i),
-                    day_of_quarter: optional_i64_at(&ov[7], i),
-                };
-                let time_overrides = LocalTimeOverrides {
-                    hour: optional_i64_at(&ov[8], i),
-                    minute: optional_i64_at(&ov[9], i),
-                    second: optional_i64_at(&ov[10], i),
-                    millisecond: optional_i64_at(&ov[11], i),
-                    microsecond: optional_i64_at(&ov[12], i),
-                    nanosecond: optional_i64_at(&ov[13], i),
-                };
-                let date = project_date(base_date(i)?, &date_overrides)?;
-                let time = project_localtime(base_time(i)?, &time_overrides)?;
-                Some((date, time))
-            })
-            .collect();
-        Ok(ColumnarValue::Array(std::sync::Arc::new(
-            build_localdatetime_struct(&parts),
-        )))
-    }
-}
-
-// ---------------------------------------------------------------------------
-// cypher_localdatetime_truncate UDF
-// ---------------------------------------------------------------------------
-
-/// `localdatetime.truncate(unit, value, map)` (`Temporal9`): truncate `value` to
-/// `unit` — for `day`-and-coarser units the date is truncated and the time zeroed
-/// to midnight; for finer units (`hour`…`microsecond`) the date is kept and the
-/// time-of-day floored — then the override `map` is applied. Args are `[value,
-/// unit, year, month, day, week, dayOfWeek, ordinalDay, quarter, dayOfQuarter,
-/// hour, minute, second, millisecond, microsecond, nanosecond]`. Returns the
-/// `localdatetime` struct. (#920)
-static CYPHER_LOCALDATETIME_TRUNCATE: LazyLock<ScalarUDF> =
-    LazyLock::new(|| ScalarUDF::new_from_impl(CypherLocalDateTimeTruncate::new()));
-
-#[derive(Debug, PartialEq, Eq, Hash)]
-struct CypherLocalDateTimeTruncate {
-    signature: Signature,
-}
-
-impl CypherLocalDateTimeTruncate {
-    fn new() -> Self {
-        Self {
-            signature: Signature::any(16, Volatility::Immutable),
-        }
-    }
-}
-
-impl ScalarUDFImpl for CypherLocalDateTimeTruncate {
-    fn name(&self) -> &'static str {
-        "cypher_localdatetime_truncate"
-    }
-
-    fn signature(&self) -> &Signature {
-        &self.signature
-    }
-
-    fn return_type(&self, _arg_types: &[DataType]) -> datafusion::error::Result<DataType> {
-        Ok(DataType::Struct(localdatetime_fields()))
-    }
-
-    #[allow(
-        clippy::too_many_lines,
-        reason = "one cohesive per-row truncation: typed source extraction, the \
-                  date/time granularity split, and 14 component overrides"
-    )]
-    fn invoke_with_args(
-        &self,
-        args: ScalarFunctionArgs,
-    ) -> datafusion::error::Result<ColumnarValue> {
-        use crate::temporal::{
-            DateOverrides, LocalTimeOverrides, parse_date_or_datetime_prefix, project_date,
-            project_localtime, time_of_day_nanos_any, truncate_date, truncate_time_nanos,
-        };
-        use datafusion::arrow::array::{
-            Array, ArrayRef, StringArray, StructArray, Time64NanosecondArray,
-        };
-        use datafusion::arrow::compute::cast;
-        use datafusion::arrow::datatypes::TimeUnit;
-        use datafusion::error::DataFusionError;
-
-        validate_heterogeneous_arguments(&args.args)?;
-        let rows = args.number_rows;
-        let cols = udf_argument_arrays(&args)?;
-        // One `value` source feeds both the date and time components.
-        let value: ArrayRef =
-            if matches!(cols[0].data_type(), DataType::Time64(TimeUnit::Nanosecond))
-                || is_date_struct(cols[0].data_type())
-                || is_localdatetime_struct(cols[0].data_type())
-                || is_time_struct(cols[0].data_type())
-                || is_datetime_struct(cols[0].data_type())
-            {
-                std::sync::Arc::clone(&cols[0])
-            } else {
-                cast(&cols[0], &DataType::Utf8).map_err(DataFusionError::from)?
-            };
-        let units_arr = cast(&cols[1], &DataType::Utf8).map_err(DataFusionError::from)?;
-        let units = units_arr.as_any().downcast_ref::<StringArray>();
-        let ov = cast_argument_arrays(&cols[2..16], &DataType::Int64)?;
-
-        let base_date = |i: usize| -> Option<i64> {
-            if value.is_null(i) {
-                return None;
-            }
-            match value.data_type() {
-                DataType::Struct(_) => {
-                    let s = value.as_any().downcast_ref::<StructArray>()?;
-                    if is_date_struct(value.data_type()) {
-                        date_struct_value(s, i)
-                    } else {
-                        Some(localdatetime_struct_parts(s, i)?.0)
-                    }
-                }
-                DataType::Utf8 => parse_date_or_datetime_prefix(
-                    value.as_any().downcast_ref::<StringArray>()?.value(i),
-                ),
-                _ => None,
-            }
-        };
-        let base_time = |i: usize| -> Option<i64> {
-            if value.is_null(i) {
-                return None;
-            }
-            match value.data_type() {
-                DataType::Time64(TimeUnit::Nanosecond) => Some(
-                    value
-                        .as_any()
-                        .downcast_ref::<Time64NanosecondArray>()?
-                        .value(i),
-                ),
-                DataType::Struct(_) => {
-                    let s = value.as_any().downcast_ref::<StructArray>()?;
-                    if is_date_struct(value.data_type()) {
-                        Some(0) // date-only → midnight
-                    } else if is_time_struct(value.data_type()) {
-                        Some(time_struct_parts(s, i)?.0)
-                    } else {
-                        Some(localdatetime_struct_parts(s, i)?.1)
-                    }
-                }
-                DataType::Utf8 => {
-                    time_of_day_nanos_any(value.as_any().downcast_ref::<StringArray>()?.value(i))
-                }
-                _ => None,
-            }
-        };
-
-        let parts: Vec<Option<(i64, i64)>> = (0..rows)
-            .map(|i| {
-                let u = units?;
-                if u.is_null(i) {
-                    return None;
-                }
-                // A `day`-and-coarser unit truncates the date and zeroes the time;
-                // a finer unit keeps the date and floors the time-of-day.
-                let (date, time) = match truncate_date(base_date(i)?, u.value(i)) {
-                    Some(d) => (d, 0i64),
-                    None => (
-                        base_date(i)?,
-                        truncate_time_nanos(base_time(i)?, u.value(i))?,
-                    ),
-                };
-                let date_overrides = DateOverrides {
-                    year: optional_i64_at(&ov[0], i),
-                    month: optional_i64_at(&ov[1], i),
-                    day: optional_i64_at(&ov[2], i),
-                    week: optional_i64_at(&ov[3], i),
-                    day_of_week: optional_i64_at(&ov[4], i),
-                    ordinal_day: optional_i64_at(&ov[5], i),
-                    quarter: optional_i64_at(&ov[6], i),
-                    day_of_quarter: optional_i64_at(&ov[7], i),
-                };
-                let time_overrides = LocalTimeOverrides {
-                    hour: optional_i64_at(&ov[8], i),
-                    minute: optional_i64_at(&ov[9], i),
-                    second: optional_i64_at(&ov[10], i),
-                    millisecond: optional_i64_at(&ov[11], i),
-                    microsecond: optional_i64_at(&ov[12], i),
-                    nanosecond: optional_i64_at(&ov[13], i),
-                };
-                let date = project_date(date, &date_overrides)?;
-                let time = project_localtime(time, &time_overrides)?;
-                Some((date, time))
-            })
-            .collect();
-        Ok(ColumnarValue::Array(std::sync::Arc::new(
-            build_localdatetime_struct(&parts),
-        )))
-    }
-}
-
-// ---------------------------------------------------------------------------
-// cypher_time_project UDF + time-struct helpers
-// ---------------------------------------------------------------------------
-
-/// The Arrow fields of a `time` value — `Struct{time: Time64(Nanosecond),
-/// offset: Int32}` (nanoseconds-of-day + zone offset in seconds).
-fn time_fields() -> datafusion::arrow::datatypes::Fields {
-    graphforge_ir::arrow_schema::time_struct_fields()
-}
-
-/// True if `dt` is the `time` struct type (dispatches base extraction/rendering
-/// without colliding with the `localdatetime` struct or user maps).
-fn is_time_struct(dt: &DataType) -> bool {
-    use datafusion::arrow::datatypes::TimeUnit;
-    matches!(dt, DataType::Struct(fields)
-        if fields.len() == 2
-            && fields[0].name() == "time"
-            && *fields[0].data_type() == DataType::Time64(TimeUnit::Nanosecond)
-            && fields[1].name() == "offset"
-            && *fields[1].data_type() == DataType::Int32)
-}
-
-/// Build a `time` struct array from per-row `(nanos_of_day, offset_seconds)`
-/// (`None` ⇒ a null row).
-fn build_time_struct(rows: &[Option<(i64, i32)>]) -> datafusion::arrow::array::StructArray {
-    use datafusion::arrow::array::{Int32Array, Time64NanosecondArray};
-    use datafusion::arrow::buffer::NullBuffer;
-    let nanos: Time64NanosecondArray = rows.iter().map(|r| r.map(|(n, _)| n)).collect();
-    let offset: Int32Array = rows.iter().map(|r| r.map(|(_, o)| o)).collect();
-    let nulls = rows.iter().map(Option::is_some).collect::<NullBuffer>();
-    datafusion::arrow::array::StructArray::new(
-        time_fields(),
-        vec![std::sync::Arc::new(nanos), std::sync::Arc::new(offset)],
-        Some(nulls),
-    )
-}
-
-/// A `time` scalar (`None` ⇒ a null value).
-fn time_scalar(parts: Option<(i64, i32)>) -> ScalarValue {
-    ScalarValue::Struct(std::sync::Arc::new(build_time_struct(&[parts])))
-}
-
-/// Extract `(nanos_of_day, offset_seconds)` from a `time` struct array at row
-/// `i` (`None` for a null row).
-fn time_struct_parts(arr: &datafusion::arrow::array::StructArray, i: usize) -> Option<(i64, i32)> {
-    use datafusion::arrow::array::{Array, Int32Array, Time64NanosecondArray};
-    if arr.is_null(i) {
-        return None;
-    }
-    let t = arr
-        .column(0)
-        .as_any()
-        .downcast_ref::<Time64NanosecondArray>()?;
-    let o = arr.column(1).as_any().downcast_ref::<Int32Array>()?;
-    (!t.is_null(i) && !o.is_null(i)).then(|| (t.value(i), o.value(i)))
-}
-
-/// `time`-from-value projection (`Temporal3`). Args are `[base, hour, minute,
-/// second, millisecond, microsecond, nanosecond, timezone]`. `base` is a
-/// `Time64`/`time`-struct/`localdatetime`-struct/temporal-string (its time-of-day
-/// and, for `time`/`datetime` bases, its offset are read); the six integer
-/// overrides adjust the time-of-day; `timezone` (a `+HH:MM`/`Z` string, or null)
-/// attaches a zone — shifting the wall-clock time if the base already had one.
-/// Returns the `time` struct. (ADR 0009)
-static CYPHER_TIME_PROJECT: LazyLock<ScalarUDF> =
-    LazyLock::new(|| ScalarUDF::new_from_impl(CypherTimeProject::new()));
-
-#[derive(Debug, PartialEq, Eq, Hash)]
-struct CypherTimeProject {
-    signature: Signature,
-}
-
-impl CypherTimeProject {
-    fn new() -> Self {
-        Self {
-            signature: Signature::any(8, Volatility::Immutable),
-        }
-    }
-}
-
-impl ScalarUDFImpl for CypherTimeProject {
-    fn name(&self) -> &'static str {
-        "cypher_time_project"
-    }
-
-    fn signature(&self) -> &Signature {
-        &self.signature
-    }
-
-    fn return_type(&self, _arg_types: &[DataType]) -> datafusion::error::Result<DataType> {
-        Ok(DataType::Struct(time_fields()))
-    }
-
-    fn invoke_with_args(
-        &self,
-        args: ScalarFunctionArgs,
-    ) -> datafusion::error::Result<ColumnarValue> {
-        use crate::temporal::{
-            LocalTimeOverrides, parse_offset_seconds, project_localtime, project_time,
-            time_of_day_with_offset,
-        };
-        use datafusion::arrow::array::{
-            Array, ArrayRef, StringArray, StructArray, Time64NanosecondArray,
-        };
-        use datafusion::arrow::compute::cast;
-        use datafusion::arrow::datatypes::TimeUnit;
-        use datafusion::error::DataFusionError;
-
-        validate_heterogeneous_arguments(&args.args)?;
-        let rows = args.number_rows;
-        let cols = udf_argument_arrays(&args)?;
-        let base: ArrayRef =
-            if matches!(cols[0].data_type(), DataType::Time64(TimeUnit::Nanosecond))
-                || is_time_struct(cols[0].data_type())
-                || is_localdatetime_struct(cols[0].data_type())
-                || is_datetime_struct(cols[0].data_type())
-            {
-                std::sync::Arc::clone(&cols[0])
-            } else {
-                cast(&cols[0], &DataType::Utf8).map_err(DataFusionError::from)?
-            };
-        let ov = cast_argument_arrays(&cols[1..7], &DataType::Int64)?;
-        let tz_arr = cast(&cols[7], &DataType::Utf8).map_err(DataFusionError::from)?;
-        let tz = tz_arr.as_any().downcast_ref::<StringArray>();
-
-        // Base time-of-day + whether it carried an offset (`None` ⇒ attach a new
-        // zone; `Some` ⇒ shift to preserve the instant).
-        let base_parts = |i: usize| -> Option<(i64, Option<i32>)> {
-            if base.is_null(i) {
-                return None;
-            }
-            match base.data_type() {
-                DataType::Time64(TimeUnit::Nanosecond) => Some((
-                    base.as_any()
-                        .downcast_ref::<Time64NanosecondArray>()?
-                        .value(i),
-                    None,
-                )),
-                DataType::Struct(_) => {
-                    let s = base.as_any().downcast_ref::<StructArray>()?;
-                    if is_time_struct(base.data_type()) {
-                        let (n, o) = time_struct_parts(s, i)?;
-                        Some((n, Some(o)))
-                    } else if is_datetime_struct(base.data_type()) {
-                        // A `datetime` carries its zone offset — keep it so a new
-                        // zone shifts the instant (`time(datetime)`).
-                        let (_, n, o, _) = datetime_struct_parts(s, i)?;
-                        Some((n, Some(o)))
-                    } else {
-                        // localdatetime struct — its time-of-day, no offset.
-                        Some((localdatetime_struct_parts(s, i)?.1, None))
-                    }
-                }
-                DataType::Utf8 => {
-                    time_of_day_with_offset(base.as_any().downcast_ref::<StringArray>()?.value(i))
-                }
-                _ => None,
-            }
-        };
-
-        let parts: Vec<Option<(i64, i32)>> = (0..rows)
-            .map(|i| {
-                let (base_nanos, base_offset) = base_parts(i)?;
-                let overrides = LocalTimeOverrides {
-                    hour: optional_i64_at(&ov[0], i),
-                    minute: optional_i64_at(&ov[1], i),
-                    second: optional_i64_at(&ov[2], i),
-                    millisecond: optional_i64_at(&ov[3], i),
-                    microsecond: optional_i64_at(&ov[4], i),
-                    nanosecond: optional_i64_at(&ov[5], i),
-                };
-                let nanos = project_localtime(base_nanos, &overrides)?;
-                // A `timezone` override (offset string) re-zones the value.
-                let new_offset = match tz {
-                    Some(a) if !a.is_null(i) => Some(parse_offset_seconds(a.value(i))?),
-                    _ => None,
-                };
-                Some(project_time(nanos, base_offset, new_offset))
-            })
-            .collect();
-        Ok(ColumnarValue::Array(std::sync::Arc::new(
-            build_time_struct(&parts),
-        )))
-    }
-}
-
-// ---------------------------------------------------------------------------
-// cypher_time_truncate UDF
-// ---------------------------------------------------------------------------
-
-/// `time.truncate(unit, value, map)` (`Temporal9`): truncate `value`'s
-/// time-of-day to `unit` (keeping its zone offset), then apply the override
-/// `map`. Args are `[value, unit, hour, minute, second, millisecond,
-/// microsecond, nanosecond, timezone]`. Returns the `time` struct. (#920)
-static CYPHER_TIME_TRUNCATE: LazyLock<ScalarUDF> =
-    LazyLock::new(|| ScalarUDF::new_from_impl(CypherTimeTruncate::new()));
-
-#[derive(Debug, PartialEq, Eq, Hash)]
-struct CypherTimeTruncate {
-    signature: Signature,
-}
-
-impl CypherTimeTruncate {
-    fn new() -> Self {
-        Self {
-            signature: Signature::any(9, Volatility::Immutable),
-        }
-    }
-}
-
-impl ScalarUDFImpl for CypherTimeTruncate {
-    fn name(&self) -> &'static str {
-        "cypher_time_truncate"
-    }
-
-    fn signature(&self) -> &Signature {
-        &self.signature
-    }
-
-    fn return_type(&self, _arg_types: &[DataType]) -> datafusion::error::Result<DataType> {
-        Ok(DataType::Struct(time_fields()))
-    }
-
-    fn invoke_with_args(
-        &self,
-        args: ScalarFunctionArgs,
-    ) -> datafusion::error::Result<ColumnarValue> {
-        use crate::temporal::{
-            LocalTimeOverrides, parse_offset_seconds, project_localtime, project_time,
-            time_of_day_with_offset, truncate_time_nanos,
-        };
-        use datafusion::arrow::array::{
-            Array, ArrayRef, StringArray, StructArray, Time64NanosecondArray,
-        };
-        use datafusion::arrow::compute::cast;
-        use datafusion::arrow::datatypes::TimeUnit;
-        use datafusion::error::DataFusionError;
-
-        validate_heterogeneous_arguments(&args.args)?;
-        let rows = args.number_rows;
-        let cols = udf_argument_arrays(&args)?;
-        let base: ArrayRef =
-            if matches!(cols[0].data_type(), DataType::Time64(TimeUnit::Nanosecond))
-                || is_time_struct(cols[0].data_type())
-                || is_localdatetime_struct(cols[0].data_type())
-                || is_datetime_struct(cols[0].data_type())
-            {
-                std::sync::Arc::clone(&cols[0])
-            } else {
-                cast(&cols[0], &DataType::Utf8).map_err(DataFusionError::from)?
-            };
-        let units_arr = cast(&cols[1], &DataType::Utf8).map_err(DataFusionError::from)?;
-        let units = units_arr.as_any().downcast_ref::<StringArray>();
-        let ov = cast_argument_arrays(&cols[2..8], &DataType::Int64)?;
-        let tz_arr = cast(&cols[8], &DataType::Utf8).map_err(DataFusionError::from)?;
-        let tz = tz_arr.as_any().downcast_ref::<StringArray>();
-
-        let base_parts = |i: usize| -> Option<(i64, Option<i32>)> {
-            if base.is_null(i) {
-                return None;
-            }
-            match base.data_type() {
-                DataType::Time64(TimeUnit::Nanosecond) => Some((
-                    base.as_any()
-                        .downcast_ref::<Time64NanosecondArray>()?
-                        .value(i),
-                    None,
-                )),
-                DataType::Struct(_) => {
-                    let s = base.as_any().downcast_ref::<StructArray>()?;
-                    if is_time_struct(base.data_type()) {
-                        let (n, o) = time_struct_parts(s, i)?;
-                        Some((n, Some(o)))
-                    } else if is_datetime_struct(base.data_type()) {
-                        let (_, n, o, _) = datetime_struct_parts(s, i)?;
-                        Some((n, Some(o)))
-                    } else {
-                        Some((localdatetime_struct_parts(s, i)?.1, None))
-                    }
-                }
-                DataType::Utf8 => {
-                    time_of_day_with_offset(base.as_any().downcast_ref::<StringArray>()?.value(i))
-                }
-                _ => None,
-            }
-        };
-
-        let parts: Vec<Option<(i64, i32)>> = (0..rows)
-            .map(|i| {
-                let u = units?;
-                if u.is_null(i) {
-                    return None;
-                }
-                let (base_nanos, base_offset) = base_parts(i)?;
-                let truncated = truncate_time_nanos(base_nanos, u.value(i))?;
-                let overrides = LocalTimeOverrides {
-                    hour: optional_i64_at(&ov[0], i),
-                    minute: optional_i64_at(&ov[1], i),
-                    second: optional_i64_at(&ov[2], i),
-                    millisecond: optional_i64_at(&ov[3], i),
-                    microsecond: optional_i64_at(&ov[4], i),
-                    nanosecond: optional_i64_at(&ov[5], i),
-                };
-                let nanos = project_localtime(truncated, &overrides)?;
-                let new_offset = match tz {
-                    Some(a) if !a.is_null(i) => Some(parse_offset_seconds(a.value(i))?),
-                    _ => None,
-                };
-                // A `timezone` override ATTACHES to the truncated wall-clock (the
-                // instant is not shifted) — pass `None` as the base offset so
-                // `project_time` attaches, mirroring datetime.truncate (#990).
-                // (#1008, Temporal9 [5])
-                let eff_offset = if new_offset.is_some() {
-                    None
-                } else {
-                    base_offset
-                };
-                Some(project_time(nanos, eff_offset, new_offset))
-            })
-            .collect();
-        Ok(ColumnarValue::Array(std::sync::Arc::new(
-            build_time_struct(&parts),
-        )))
-    }
-}
-
-// ---------------------------------------------------------------------------
-// cypher_datetime_project UDF + datetime-struct helpers
-// ---------------------------------------------------------------------------
-
-/// One `datetime` row: `(date_days, nanos_of_day, offset_seconds, zone_label)`,
-/// or `None` for a null value.
-type DateTimeRow = Option<(i64, i64, i32, Option<String>)>;
-
-/// The Arrow fields of a `datetime` value — `Struct{date: Int64, time:
-/// Time64(Nanosecond), offset: Int32, zone: Utf8}` (the date = i64 days,
-/// time-of-day, resolved zone offset in seconds, and an optional named-IANA-zone
-/// label). (#1011)
-fn datetime_fields() -> datafusion::arrow::datatypes::Fields {
-    graphforge_ir::arrow_schema::datetime_struct_fields()
-}
-
-/// True if `dt` is the `datetime` struct type.
-fn is_datetime_struct(dt: &DataType) -> bool {
-    use datafusion::arrow::datatypes::TimeUnit;
-    matches!(dt, DataType::Struct(fields)
-        if fields.len() == 4
-            && fields[0].name() == "date" && *fields[0].data_type() == DataType::Int64
-            && fields[1].name() == "time"
-            && *fields[1].data_type() == DataType::Time64(TimeUnit::Nanosecond)
-            && fields[2].name() == "offset" && *fields[2].data_type() == DataType::Int32
-            && fields[3].name() == "zone" && *fields[3].data_type() == DataType::Utf8)
-}
-
-/// Build a `datetime` struct array from per-row `(date_days, nanos_of_day,
-/// offset_seconds, zone_label)` (`None` ⇒ a null row).
-fn build_datetime_struct(rows: &[DateTimeRow]) -> datafusion::arrow::array::StructArray {
-    use datafusion::arrow::array::{Int32Array, Int64Array, StringArray, Time64NanosecondArray};
-    use datafusion::arrow::buffer::NullBuffer;
-    let days: Int64Array = rows.iter().map(|r| r.as_ref().map(|t| t.0)).collect();
-    let nanos: Time64NanosecondArray = rows.iter().map(|r| r.as_ref().map(|t| t.1)).collect();
-    let offset: Int32Array = rows.iter().map(|r| r.as_ref().map(|t| t.2)).collect();
-    // The zone field is empty (NOT null) when there is no named zone, so two
-    // offset-only datetimes compare equal — `cypher_struct_eq` propagates null,
-    // and a null=null field would make the whole equality null (Temporal7 [5]).
-    let zone: StringArray = rows
-        .iter()
-        .map(|r| r.as_ref().map(|t| t.3.clone().unwrap_or_default()))
-        .collect();
-    let nulls = rows.iter().map(Option::is_some).collect::<NullBuffer>();
-    datafusion::arrow::array::StructArray::new(
-        datetime_fields(),
-        vec![
-            std::sync::Arc::new(days),
-            std::sync::Arc::new(nanos),
-            std::sync::Arc::new(offset),
-            std::sync::Arc::new(zone),
-        ],
-        Some(nulls),
-    )
-}
-
-/// A `datetime` scalar (`None` ⇒ a null value).
-fn datetime_scalar(parts: DateTimeRow) -> ScalarValue {
-    ScalarValue::Struct(std::sync::Arc::new(build_datetime_struct(&[parts])))
-}
-
-/// Extract `(date_days, nanos_of_day, offset_seconds, zone_label)` from a
-/// `datetime` struct array at row `i` (`None` for a null row).
-fn datetime_struct_parts(arr: &datafusion::arrow::array::StructArray, i: usize) -> DateTimeRow {
-    use datafusion::arrow::array::{
-        Array, Int32Array, Int64Array, StringArray, Time64NanosecondArray,
-    };
-    if arr.is_null(i) {
-        return None;
-    }
-    let days = arr.column(0).as_any().downcast_ref::<Int64Array>()?;
-    let nanos = arr
-        .column(1)
-        .as_any()
-        .downcast_ref::<Time64NanosecondArray>()?;
-    let offset = arr.column(2).as_any().downcast_ref::<Int32Array>()?;
-    let zone = arr.column(3).as_any().downcast_ref::<StringArray>()?;
-    if days.is_null(i) || nanos.is_null(i) || offset.is_null(i) {
-        return None;
-    }
-    // An empty zone label means "no named zone" (offset-only datetime).
-    let zone_label =
-        (!zone.is_null(i) && !zone.value(i).is_empty()).then(|| zone.value(i).to_string());
-    Some((days.value(i), nanos.value(i), offset.value(i), zone_label))
-}
-
-/// `datetime`-from-value projection (`Temporal3` [8]-[11]). Args are `[date_src,
-/// time_src, year, month, day, week, dayOfWeek, ordinalDay, quarter,
-/// dayOfQuarter, hour, minute, second, millisecond, microsecond, nanosecond,
-/// timezone]`. The date/time sources are any temporal value/string (the time
-/// source also carries the source offset + named zone); the 14 integer overrides
-/// adjust the local date/time; `timezone` re-zones. Returns the `datetime`
-/// struct. (ADR 0009)
-static CYPHER_DATETIME_PROJECT: LazyLock<ScalarUDF> =
-    LazyLock::new(|| ScalarUDF::new_from_impl(CypherDateTimeProject::new()));
-
-#[derive(Debug, PartialEq, Eq, Hash)]
-struct CypherDateTimeProject {
-    signature: Signature,
-}
-
-impl CypherDateTimeProject {
-    fn new() -> Self {
-        Self {
-            signature: Signature::any(17, Volatility::Immutable),
-        }
-    }
-}
-
-impl ScalarUDFImpl for CypherDateTimeProject {
-    fn name(&self) -> &'static str {
-        "cypher_datetime_project"
-    }
-
-    fn signature(&self) -> &Signature {
-        &self.signature
-    }
-
-    fn return_type(&self, _arg_types: &[DataType]) -> datafusion::error::Result<DataType> {
-        Ok(DataType::Struct(datetime_fields()))
-    }
-
-    #[allow(
-        clippy::too_many_lines,
-        reason = "one cohesive per-row projection: typed date/time/zone source \
-                  extraction, 14 component overrides, and zone re-resolution"
-    )]
-    fn invoke_with_args(
-        &self,
-        args: ScalarFunctionArgs,
-    ) -> datafusion::error::Result<ColumnarValue> {
-        use crate::temporal::{
-            DateOverrides, LocalTimeOverrides, parse_date_or_datetime_prefix, project_date,
-            project_datetime, project_localtime, time_offset_zone,
-        };
-        use datafusion::arrow::array::{
-            Array, ArrayRef, StringArray, StructArray, Time64NanosecondArray,
-        };
-        use datafusion::arrow::compute::cast;
-        use datafusion::arrow::datatypes::TimeUnit;
-        use datafusion::error::DataFusionError;
-
-        validate_heterogeneous_arguments(&args.args)?;
-        let rows = args.number_rows;
-        let cols = udf_argument_arrays(&args)?;
-        let typed_or_utf8 = |a: &ArrayRef| -> datafusion::error::Result<ArrayRef> {
-            if matches!(a.data_type(), DataType::Time64(TimeUnit::Nanosecond))
-                || is_date_struct(a.data_type())
-                || is_localdatetime_struct(a.data_type())
-                || is_time_struct(a.data_type())
-                || is_datetime_struct(a.data_type())
-            {
-                Ok(std::sync::Arc::clone(a))
-            } else {
-                cast(a, &DataType::Utf8).map_err(DataFusionError::from)
-            }
-        };
-        let date_src = typed_or_utf8(&cols[0])?;
-        let time_src = typed_or_utf8(&cols[1])?;
-        let ov = cast_argument_arrays(&cols[2..16], &DataType::Int64)?;
-        let tz_arr = cast(&cols[16], &DataType::Utf8).map_err(DataFusionError::from)?;
-        let tz = tz_arr.as_any().downcast_ref::<StringArray>();
-
-        let base_date = |i: usize| -> Option<i64> {
-            if date_src.is_null(i) {
-                return Some(0); // a missing date defaults to the epoch (day 0)
-            }
-            match date_src.data_type() {
-                DataType::Struct(_) => {
-                    let s = date_src.as_any().downcast_ref::<StructArray>()?;
-                    if is_date_struct(date_src.data_type()) {
-                        date_struct_value(s, i)
-                    } else if is_datetime_struct(date_src.data_type()) {
-                        Some(datetime_struct_parts(s, i)?.0)
-                    } else {
-                        Some(localdatetime_struct_parts(s, i)?.0)
-                    }
-                }
-                DataType::Utf8 => parse_date_or_datetime_prefix(
-                    date_src.as_any().downcast_ref::<StringArray>()?.value(i),
-                ),
-                _ => None,
-            }
-        };
-        // Base time-of-day plus the source's offset and named zone (if any).
-        let base_time = |i: usize| -> Option<(i64, Option<i32>, Option<String>)> {
-            if time_src.is_null(i) {
-                return Some((0, None, None));
-            }
-            match time_src.data_type() {
-                DataType::Time64(TimeUnit::Nanosecond) => Some((
-                    time_src
-                        .as_any()
-                        .downcast_ref::<Time64NanosecondArray>()?
-                        .value(i),
-                    None,
-                    None,
-                )),
-                DataType::Struct(_) => {
-                    let s = time_src.as_any().downcast_ref::<StructArray>()?;
-                    if is_date_struct(time_src.data_type()) {
-                        // A bare date carries no time-of-day (matches the pre-#1011
-                        // `Date32` fall-through: `datetime(date(…))` → null).
-                        None
-                    } else if is_datetime_struct(time_src.data_type()) {
-                        let (_, n, o, z) = datetime_struct_parts(s, i)?;
-                        Some((n, Some(o), z))
-                    } else if is_time_struct(time_src.data_type()) {
-                        let (n, o) = time_struct_parts(s, i)?;
-                        Some((n, Some(o), None))
-                    } else {
-                        Some((localdatetime_struct_parts(s, i)?.1, None, None))
-                    }
-                }
-                DataType::Utf8 => {
-                    time_offset_zone(time_src.as_any().downcast_ref::<StringArray>()?.value(i))
-                }
-                _ => None,
-            }
-        };
-
-        let parts: Vec<DateTimeRow> = (0..rows)
-            .map(|i| {
-                let date_overrides = DateOverrides {
-                    year: optional_i64_at(&ov[0], i),
-                    month: optional_i64_at(&ov[1], i),
-                    day: optional_i64_at(&ov[2], i),
-                    week: optional_i64_at(&ov[3], i),
-                    day_of_week: optional_i64_at(&ov[4], i),
-                    ordinal_day: optional_i64_at(&ov[5], i),
-                    quarter: optional_i64_at(&ov[6], i),
-                    day_of_quarter: optional_i64_at(&ov[7], i),
-                };
-                let time_overrides = LocalTimeOverrides {
-                    hour: optional_i64_at(&ov[8], i),
-                    minute: optional_i64_at(&ov[9], i),
-                    second: optional_i64_at(&ov[10], i),
-                    millisecond: optional_i64_at(&ov[11], i),
-                    microsecond: optional_i64_at(&ov[12], i),
-                    nanosecond: optional_i64_at(&ov[13], i),
-                };
-                let (base_nanos, src_offset, src_zone) = base_time(i)?;
-                let date = project_date(base_date(i)?, &date_overrides)?;
-                let nanos = project_localtime(base_nanos, &time_overrides)?;
-                let new_tz = tz.and_then(|a| (!a.is_null(i)).then(|| a.value(i)));
-                let (date, nanos, offset, zone) =
-                    project_datetime(date, nanos, src_offset, src_zone.as_deref(), new_tz)?;
-                Some((date, nanos, offset, zone))
-            })
-            .collect();
-        Ok(ColumnarValue::Array(std::sync::Arc::new(
-            build_datetime_struct(&parts),
-        )))
-    }
-}
-
-// ---------------------------------------------------------------------------
-// cypher_datetime_truncate UDF
-// ---------------------------------------------------------------------------
-
-/// `datetime.truncate(unit, value, map)` (`Temporal9`): truncate `value` to
-/// `unit` — date for `day`-and-coarser units (time zeroed), time-of-day for finer
-/// units — keeping the source zone, then apply the override `map` and optional
-/// `timezone`. Args are `[value, unit, year, month, day, week, dayOfWeek,
-/// ordinalDay, quarter, dayOfQuarter, hour, minute, second, millisecond,
-/// microsecond, nanosecond, timezone]`. Returns the `datetime` struct. (#920)
-static CYPHER_DATETIME_TRUNCATE: LazyLock<ScalarUDF> =
-    LazyLock::new(|| ScalarUDF::new_from_impl(CypherDateTimeTruncate::new()));
-
-#[derive(Debug, PartialEq, Eq, Hash)]
-struct CypherDateTimeTruncate {
-    signature: Signature,
-}
-
-impl CypherDateTimeTruncate {
-    fn new() -> Self {
-        Self {
-            signature: Signature::any(17, Volatility::Immutable),
-        }
-    }
-}
-
-impl ScalarUDFImpl for CypherDateTimeTruncate {
-    fn name(&self) -> &'static str {
-        "cypher_datetime_truncate"
-    }
-
-    fn signature(&self) -> &Signature {
-        &self.signature
-    }
-
-    fn return_type(&self, _arg_types: &[DataType]) -> datafusion::error::Result<DataType> {
-        Ok(DataType::Struct(datetime_fields()))
-    }
-
-    #[allow(
-        clippy::too_many_lines,
-        reason = "one cohesive per-row truncation: typed source extraction, the \
-                  date/time granularity split, 14 overrides, and zone re-resolution"
-    )]
-    fn invoke_with_args(
-        &self,
-        args: ScalarFunctionArgs,
-    ) -> datafusion::error::Result<ColumnarValue> {
-        use crate::temporal::{
-            DateOverrides, LocalTimeOverrides, parse_date_or_datetime_prefix, project_date,
-            project_datetime, project_localtime, time_offset_zone, truncate_date,
-            truncate_time_nanos,
-        };
-        use datafusion::arrow::array::{
-            Array, ArrayRef, StringArray, StructArray, Time64NanosecondArray,
-        };
-        use datafusion::arrow::compute::cast;
-        use datafusion::arrow::datatypes::TimeUnit;
-        use datafusion::error::DataFusionError;
-
-        validate_heterogeneous_arguments(&args.args)?;
-        let rows = args.number_rows;
-        let cols = udf_argument_arrays(&args)?;
-        // One `value` source feeds both the date and time components.
-        let value: ArrayRef =
-            if matches!(cols[0].data_type(), DataType::Time64(TimeUnit::Nanosecond))
-                || is_date_struct(cols[0].data_type())
-                || is_localdatetime_struct(cols[0].data_type())
-                || is_time_struct(cols[0].data_type())
-                || is_datetime_struct(cols[0].data_type())
-            {
-                std::sync::Arc::clone(&cols[0])
-            } else {
-                cast(&cols[0], &DataType::Utf8).map_err(DataFusionError::from)?
-            };
-        let units_arr = cast(&cols[1], &DataType::Utf8).map_err(DataFusionError::from)?;
-        let units = units_arr.as_any().downcast_ref::<StringArray>();
-        let ov = cast_argument_arrays(&cols[2..16], &DataType::Int64)?;
-        let tz_arr = cast(&cols[16], &DataType::Utf8).map_err(DataFusionError::from)?;
-        let tz = tz_arr.as_any().downcast_ref::<StringArray>();
-
-        let base_date = |i: usize| -> Option<i64> {
-            if value.is_null(i) {
-                return None;
-            }
-            match value.data_type() {
-                DataType::Struct(_) => {
-                    let s = value.as_any().downcast_ref::<StructArray>()?;
-                    if is_date_struct(value.data_type()) {
-                        date_struct_value(s, i)
-                    } else if is_datetime_struct(value.data_type()) {
-                        Some(datetime_struct_parts(s, i)?.0)
-                    } else {
-                        Some(localdatetime_struct_parts(s, i)?.0)
-                    }
-                }
-                DataType::Utf8 => parse_date_or_datetime_prefix(
-                    value.as_any().downcast_ref::<StringArray>()?.value(i),
-                ),
-                _ => None,
-            }
-        };
-        // Base time-of-day plus the source's offset and named zone (if any).
-        let base_time = |i: usize| -> Option<(i64, Option<i32>, Option<String>)> {
-            if value.is_null(i) {
-                return None;
-            }
-            match value.data_type() {
-                DataType::Time64(TimeUnit::Nanosecond) => Some((
-                    value
-                        .as_any()
-                        .downcast_ref::<Time64NanosecondArray>()?
-                        .value(i),
-                    None,
-                    None,
-                )),
-                DataType::Struct(_) => {
-                    let s = value.as_any().downcast_ref::<StructArray>()?;
-                    if is_date_struct(value.data_type()) {
-                        Some((0, None, None)) // date-only → midnight
-                    } else if is_datetime_struct(value.data_type()) {
-                        let (_, n, o, z) = datetime_struct_parts(s, i)?;
-                        Some((n, Some(o), z))
-                    } else if is_time_struct(value.data_type()) {
-                        let (n, o) = time_struct_parts(s, i)?;
-                        Some((n, Some(o), None))
-                    } else {
-                        Some((localdatetime_struct_parts(s, i)?.1, None, None))
-                    }
-                }
-                DataType::Utf8 => {
-                    time_offset_zone(value.as_any().downcast_ref::<StringArray>()?.value(i))
-                }
-                _ => None,
-            }
-        };
-
-        let parts: Vec<DateTimeRow> = (0..rows)
-            .map(|i| {
-                let u = units?;
-                if u.is_null(i) {
-                    return None;
-                }
-                let (bt_nanos, src_offset, src_zone) = base_time(i)?;
-                // A `day`-and-coarser unit truncates the date and zeroes the time;
-                // a finer unit keeps the date and floors the time-of-day.
-                let (date0, nanos0) = match truncate_date(base_date(i)?, u.value(i)) {
-                    Some(d) => (d, 0i64),
-                    None => (base_date(i)?, truncate_time_nanos(bt_nanos, u.value(i))?),
-                };
-                let date_overrides = DateOverrides {
-                    year: optional_i64_at(&ov[0], i),
-                    month: optional_i64_at(&ov[1], i),
-                    day: optional_i64_at(&ov[2], i),
-                    week: optional_i64_at(&ov[3], i),
-                    day_of_week: optional_i64_at(&ov[4], i),
-                    ordinal_day: optional_i64_at(&ov[5], i),
-                    quarter: optional_i64_at(&ov[6], i),
-                    day_of_quarter: optional_i64_at(&ov[7], i),
-                };
-                let time_overrides = LocalTimeOverrides {
-                    hour: optional_i64_at(&ov[8], i),
-                    minute: optional_i64_at(&ov[9], i),
-                    second: optional_i64_at(&ov[10], i),
-                    millisecond: optional_i64_at(&ov[11], i),
-                    microsecond: optional_i64_at(&ov[12], i),
-                    nanosecond: optional_i64_at(&ov[13], i),
-                };
-                let date = project_date(date0, &date_overrides)?;
-                let nanos = project_localtime(nanos0, &time_overrides)?;
-                let new_tz = tz.and_then(|a| (!a.is_null(i)).then(|| a.value(i)));
-                // Truncation is WALL-CLOCK preserving: a `{timezone: …}` override
-                // ATTACHES the zone to the truncated local time (midnight stays
-                // midnight), it does not re-express the source instant. So drop
-                // the source offset when a new zone is given, forcing
-                // `project_datetime`'s attach path instead of an instant shift
-                // (#920 — otherwise `truncate(…, {timezone: 'Europe/Stockholm'})`
-                // shifted midnight by the source offset).
-                let src_offset = if new_tz.is_some() { None } else { src_offset };
-                let (date, nanos, offset, zone) =
-                    project_datetime(date, nanos, src_offset, src_zone.as_deref(), new_tz)?;
-                Some((date, nanos, offset, zone))
-            })
-            .collect();
-        Ok(ColumnarValue::Array(std::sync::Arc::new(
-            build_datetime_struct(&parts),
-        )))
-    }
-}
-
-// ---------------------------------------------------------------------------
 // cypher_to_string UDF
 // ---------------------------------------------------------------------------
 
@@ -10234,124 +7733,6 @@ impl ScalarUDFImpl for CypherToString {
             ),
         };
         Ok(result)
-    }
-}
-
-// ---------------------------------------------------------------------------
-// cypher_date_truncate UDF
-// ---------------------------------------------------------------------------
-
-/// `date.truncate(unit, value, map)` (`Temporal9`): truncate `value`'s date to
-/// `unit`, then apply the override `map` (same fields as projection). Args are
-/// `[value, unit, year, month, day, week, dayOfWeek, ordinalDay, quarter,
-/// dayOfQuarter]` — `value` is a `Date32` or ISO date/datetime string, `unit` a
-/// string, the eight overrides nullable integers. Returns `Date32`. (#920)
-static CYPHER_DATE_TRUNCATE: LazyLock<ScalarUDF> =
-    LazyLock::new(|| ScalarUDF::new_from_impl(CypherDateTruncate::new()));
-
-#[derive(Debug, PartialEq, Eq, Hash)]
-struct CypherDateTruncate {
-    signature: Signature,
-}
-
-impl CypherDateTruncate {
-    fn new() -> Self {
-        Self {
-            signature: Signature::any(10, Volatility::Immutable),
-        }
-    }
-}
-
-impl ScalarUDFImpl for CypherDateTruncate {
-    fn name(&self) -> &'static str {
-        "cypher_date_truncate"
-    }
-
-    fn signature(&self) -> &Signature {
-        &self.signature
-    }
-
-    fn return_type(&self, _arg_types: &[DataType]) -> datafusion::error::Result<DataType> {
-        Ok(DataType::Struct(
-            graphforge_ir::arrow_schema::date_struct_fields(),
-        ))
-    }
-
-    fn invoke_with_args(
-        &self,
-        args: ScalarFunctionArgs,
-    ) -> datafusion::error::Result<ColumnarValue> {
-        use crate::temporal::{
-            DateOverrides, parse_date_or_datetime_prefix, project_date, truncate_date,
-        };
-        use datafusion::arrow::array::{Array, ArrayRef, StringArray, StructArray};
-        use datafusion::arrow::compute::cast;
-        use datafusion::error::DataFusionError;
-
-        validate_heterogeneous_arguments(&args.args)?;
-        let rows = args.number_rows;
-        let cols = udf_argument_arrays(&args)?;
-        // DataFusion emits `Utf8View` for string *columns* by default (string
-        // literals stay `Utf8`), and our downcasts target `StringArray` — cast
-        // string inputs to `Utf8` so a column-typed value/unit isn't silently
-        // nulled. A `date`/`localdatetime`/`time`/`datetime` struct is taken directly.
-        let value: ArrayRef = if is_date_struct(cols[0].data_type())
-            || is_localdatetime_struct(cols[0].data_type())
-            || is_time_struct(cols[0].data_type())
-            || is_datetime_struct(cols[0].data_type())
-        {
-            std::sync::Arc::clone(&cols[0])
-        } else {
-            cast(&cols[0], &DataType::Utf8).map_err(DataFusionError::from)?
-        };
-        let units_arr = cast(&cols[1], &DataType::Utf8).map_err(DataFusionError::from)?;
-        let units = units_arr.as_any().downcast_ref::<StringArray>();
-        let ov = cast_argument_arrays(&cols[2..10], &DataType::Int64)?;
-
-        let base_date = |i: usize| -> Option<i64> {
-            if value.is_null(i) {
-                return None;
-            }
-            match value.data_type() {
-                DataType::Struct(_) => {
-                    let s = value.as_any().downcast_ref::<StructArray>()?;
-                    if is_date_struct(value.data_type()) {
-                        date_struct_value(s, i)
-                    } else {
-                        // A `localdatetime`/`datetime` value — truncate its date.
-                        Some(localdatetime_struct_parts(s, i)?.0)
-                    }
-                }
-                DataType::Utf8 => parse_date_or_datetime_prefix(
-                    value.as_any().downcast_ref::<StringArray>()?.value(i),
-                ),
-                _ => None,
-            }
-        };
-
-        let out: Vec<Option<i64>> = (0..rows)
-            .map(|i| {
-                let u = units?;
-                if u.is_null(i) {
-                    return None;
-                }
-                let truncated = truncate_date(base_date(i)?, u.value(i))?;
-                let overrides = DateOverrides {
-                    year: optional_i64_at(&ov[0], i),
-                    month: optional_i64_at(&ov[1], i),
-                    day: optional_i64_at(&ov[2], i),
-                    week: optional_i64_at(&ov[3], i),
-                    day_of_week: optional_i64_at(&ov[4], i),
-                    ordinal_day: optional_i64_at(&ov[5], i),
-                    quarter: optional_i64_at(&ov[6], i),
-                    day_of_quarter: optional_i64_at(&ov[7], i),
-                };
-                project_date(truncated, &overrides)
-            })
-            .collect();
-        Ok(ColumnarValue::Array(std::sync::Arc::new(
-            build_date_struct(&out),
-        )))
     }
 }
 
@@ -10750,226 +8131,6 @@ mod tests {
     }
 
     #[test]
-    fn temporal_clock_fn_recognition() {
-        // Every instant type × clock accessor is recognised (#920).
-        for base in ["date", "localtime", "time", "localdatetime", "datetime"] {
-            for clock in ["transaction", "statement", "realtime"] {
-                assert!(is_temporal_clock_fn(&format!("{base}.{clock}")));
-            }
-        }
-        // Case-insensitive (Cypher function names are).
-        assert!(is_temporal_clock_fn("Date.Realtime"));
-        assert!(is_temporal_clock_fn("DATETIME.TRANSACTION"));
-        // Non-clock and non-temporal names are not.
-        assert!(!is_temporal_clock_fn("datetime.truncate"));
-        assert!(!is_temporal_clock_fn("duration.realtime")); // duration has no clock
-        assert!(!is_temporal_clock_fn("date"));
-        assert!(!is_temporal_clock_fn("foo.realtime"));
-
-        // A null temporal arg lowers to a TYPED null, not generic Null.
-        assert_eq!(
-            temporal_null_scalar("date").data_type(),
-            date_scalar(None).data_type()
-        );
-        assert_eq!(
-            temporal_null_scalar("localtime"),
-            ScalarValue::Time64Nanosecond(None)
-        );
-        assert_eq!(
-            temporal_null_scalar("datetime.realtime").data_type(),
-            datetime_scalar(None).data_type()
-        );
-        assert_eq!(temporal_null_scalar("unknown"), ScalarValue::Null);
-    }
-
-    #[test]
-    fn temporal_struct_builders_and_extractors_preserve_values_and_nulls() {
-        use crate::temporal::DurationValue;
-        use datafusion::arrow::array::{ArrayRef, Int32Array, StructArray};
-        use datafusion::arrow::datatypes::{Field, Fields};
-
-        let dates = build_date_struct(&[Some(19_723), None]);
-        assert_eq!(date_struct_value(&dates, 0), Some(19_723));
-        assert_eq!(date_struct_value(&dates, 1), None);
-
-        let local = build_localdatetime_struct(&[(Some((19_723, 45_000))), None]);
-        assert_eq!(
-            localdatetime_struct_parts(&local, 0),
-            Some((19_723, 45_000))
-        );
-        assert_eq!(localdatetime_struct_parts(&local, 1), None);
-
-        let duration = DurationValue {
-            months: 14,
-            days: -2,
-            seconds: 90,
-            nanos: 123,
-        };
-        let durations = build_duration_struct(&[Some(duration), None]);
-        assert_eq!(duration_struct_parts(&durations, 0), Some(duration));
-        assert_eq!(duration_struct_parts(&durations, 1), None);
-        assert_eq!(
-            dur_secs_nanos(-3, 7),
-            DurationValue {
-                months: 0,
-                days: 0,
-                seconds: -3,
-                nanos: 7,
-            }
-        );
-        assert_eq!(
-            duration_value_to_ir(duration),
-            IrLiteral::Duration {
-                months: 14,
-                days: -2,
-                seconds: 90,
-                nanos: 123,
-            }
-        );
-
-        let times = build_time_struct(&[Some((86_399_000_000_000, -25_200)), None]);
-        assert_eq!(
-            time_struct_parts(&times, 0),
-            Some((86_399_000_000_000, -25_200))
-        );
-        assert_eq!(time_struct_parts(&times, 1), None);
-
-        let datetimes = build_datetime_struct(&[
-            Some((19_723, 45_000, 3_600, Some("Europe/Paris".into()))),
-            Some((19_724, 46_000, 0, None)),
-            None,
-        ]);
-        assert_eq!(
-            datetime_struct_parts(&datetimes, 0),
-            Some((19_723, 45_000, 3_600, Some("Europe/Paris".into())))
-        );
-        assert_eq!(
-            datetime_struct_parts(&datetimes, 1),
-            Some((19_724, 46_000, 0, None))
-        );
-        assert_eq!(datetime_struct_parts(&datetimes, 2), None);
-
-        // Extractors reject a structurally wrong child type without fabricating
-        // a value. This exercises the defensive downcast path with valid Arrow.
-        let wrong_fields = Fields::from(vec![Field::new("epoch_day", DataType::Int32, true)]);
-        let wrong_children: Vec<ArrayRef> = vec![Arc::new(Int32Array::from(vec![Some(1)]))];
-        let wrong_date = StructArray::new(wrong_fields, wrong_children, None);
-        assert_eq!(date_struct_value(&wrong_date, 0), None);
-
-        let overrides: ArrayRef = Arc::new(datafusion::arrow::array::Int64Array::from(vec![
-            Some(8),
-            None,
-        ]));
-        assert_eq!(optional_i64_at(&overrides, 0), Some(8));
-        assert_eq!(optional_i64_at(&overrides, 1), None);
-        let wrong_override: ArrayRef = Arc::new(Int32Array::from(vec![Some(8)]));
-        assert_eq!(optional_i64_at(&wrong_override, 0), None);
-    }
-
-    #[test]
-    fn temporal_project_and_truncate_udfs_execute_all_typed_families() {
-        use datafusion::arrow::array::Array;
-
-        let null_ints = |count| vec![ScalarValue::Int64(None); count];
-        let assert_value = |array: datafusion::arrow::array::ArrayRef| {
-            assert_eq!(array.len(), 1);
-            assert!(!array.is_null(0));
-        };
-
-        let mut args = vec![ScalarValue::Utf8(Some("2024-02-29".into()))];
-        args.extend(null_ints(8));
-        assert_value(invoke_test_udf(&CypherDateProject::new(), args).unwrap());
-        let mut null_args = vec![ScalarValue::Utf8(None)];
-        null_args.extend(null_ints(8));
-        let null_date = invoke_test_udf(&CypherDateProject::new(), null_args).unwrap();
-        assert!(null_date.is_null(0));
-        let mut typed_args = vec![date_scalar(Some(19_782))];
-        typed_args.extend(null_ints(8));
-        assert_value(invoke_test_udf(&CypherDateProject::new(), typed_args).unwrap());
-
-        let mut args = vec![ScalarValue::Utf8(Some("12:34:56.123".into()))];
-        args.extend(null_ints(6));
-        assert_value(invoke_test_udf(&CypherLocalTimeProject::new(), args).unwrap());
-        let mut typed_args = vec![ScalarValue::Time64Nanosecond(Some(45_296_123_000_000))];
-        typed_args.extend(null_ints(6));
-        assert_value(invoke_test_udf(&CypherLocalTimeProject::new(), typed_args).unwrap());
-
-        let mut args = vec![
-            ScalarValue::Utf8(Some("12:34:56.123".into())),
-            ScalarValue::Utf8(Some("second".into())),
-        ];
-        args.extend(null_ints(6));
-        assert_value(invoke_test_udf(&CypherLocalTimeTruncate::new(), args).unwrap());
-
-        let mut args = vec![
-            ScalarValue::Utf8(Some("2024-02-29".into())),
-            ScalarValue::Utf8(Some("12:34:56.123".into())),
-        ];
-        args.extend(null_ints(14));
-        assert_value(invoke_test_udf(&CypherLocalDateTimeProject::new(), args).unwrap());
-        let mut typed_args = vec![
-            date_scalar(Some(19_782)),
-            ScalarValue::Time64Nanosecond(Some(45_296_123_000_000)),
-        ];
-        typed_args.extend(null_ints(14));
-        assert_value(invoke_test_udf(&CypherLocalDateTimeProject::new(), typed_args).unwrap());
-
-        let mut args = vec![
-            ScalarValue::Utf8(Some("2024-02-29T12:34:56.123".into())),
-            ScalarValue::Utf8(Some("day".into())),
-        ];
-        args.extend(null_ints(14));
-        assert_value(invoke_test_udf(&CypherLocalDateTimeTruncate::new(), args).unwrap());
-
-        let mut args = vec![ScalarValue::Utf8(Some("12:34:56+01:00".into()))];
-        args.extend(null_ints(6));
-        args.push(ScalarValue::Utf8(None));
-        assert_value(invoke_test_udf(&CypherTimeProject::new(), args).unwrap());
-        let mut typed_args = vec![time_scalar(Some((45_296_000_000_000, 3_600)))];
-        typed_args.extend(null_ints(6));
-        typed_args.push(ScalarValue::Utf8(None));
-        assert_value(invoke_test_udf(&CypherTimeProject::new(), typed_args).unwrap());
-
-        let mut args = vec![
-            ScalarValue::Utf8(Some("12:34:56+01:00".into())),
-            ScalarValue::Utf8(Some("minute".into())),
-        ];
-        args.extend(null_ints(6));
-        args.push(ScalarValue::Utf8(None));
-        assert_value(invoke_test_udf(&CypherTimeTruncate::new(), args).unwrap());
-
-        let mut args = vec![
-            ScalarValue::Utf8(Some("2024-02-29".into())),
-            ScalarValue::Utf8(Some("12:34:56+01:00".into())),
-        ];
-        args.extend(null_ints(14));
-        args.push(ScalarValue::Utf8(None));
-        assert_value(invoke_test_udf(&CypherDateTimeProject::new(), args).unwrap());
-        let mut typed_args = vec![
-            date_scalar(Some(19_782)),
-            time_scalar(Some((45_296_000_000_000, 3_600))),
-        ];
-        typed_args.extend(null_ints(14));
-        typed_args.push(ScalarValue::Utf8(None));
-        assert_value(invoke_test_udf(&CypherDateTimeProject::new(), typed_args).unwrap());
-
-        let mut args = vec![
-            ScalarValue::Utf8(Some("2024-02-29T12:34:56+01:00".into())),
-            ScalarValue::Utf8(Some("hour".into())),
-        ];
-        args.extend(null_ints(14));
-        args.push(ScalarValue::Utf8(None));
-        assert_value(invoke_test_udf(&CypherDateTimeTruncate::new(), args).unwrap());
-
-        let mut args = vec![
-            ScalarValue::Utf8(Some("2024-02-29".into())),
-            ScalarValue::Utf8(Some("month".into())),
-        ];
-        args.extend(null_ints(8));
-        assert_value(invoke_test_udf(&CypherDateTruncate::new(), args).unwrap());
-    }
-
-    #[test]
     fn quantifier_three_valued_reduce() {
         use datafusion::arrow::array::BooleanArray;
         use graphforge_ir::QuantifierKind::{All, Any, None, Single};
@@ -11209,36 +8370,6 @@ mod tests {
                 .collect::<Vec<_>>(),
             vec![ScalarValue::Int64(Some(5)), ScalarValue::Int64(Some(7))]
         );
-    }
-
-    #[test]
-    fn shared_temporal_cast_helper_preserves_arrow_values_nulls_and_error() {
-        use datafusion::arrow::array::{Array, ArrayRef, Int32Array, Int64Array};
-
-        let integers: ArrayRef = Arc::new(Int32Array::from(vec![Some(7), None]));
-        let casted = cast_argument_arrays(&[integers], &DataType::Int64).unwrap();
-        let casted = casted[0]
-            .as_any()
-            .downcast_ref::<Int64Array>()
-            .expect("Int64");
-        assert_eq!(casted.value(0), 7);
-        assert!(casted.is_null(1));
-
-        let invalid = ScalarValue::List(ScalarValue::new_list(
-            &[ScalarValue::Int64(Some(1))],
-            &DataType::Int64,
-            true,
-        ))
-        .to_array_of_size(1)
-        .unwrap();
-        let direct_error = datafusion::arrow::compute::cast(&invalid, &DataType::Int64)
-            .map_err(datafusion::error::DataFusionError::from)
-            .unwrap_err()
-            .to_string();
-        let helper_error = cast_argument_arrays(&[invalid], &DataType::Int64)
-            .unwrap_err()
-            .to_string();
-        assert_eq!(helper_error, direct_error);
     }
 
     #[test]
@@ -13409,31 +10540,6 @@ mod tests {
     }
 
     #[test]
-    fn temporal_accessor_type_matrix_distinguishes_values_from_properties() {
-        // Date and duration dispatch through their dedicated lowering paths.
-        assert!(!temporal_accessor_valid(&DataType::Date32, "year"));
-        assert!(!temporal_accessor_valid(&DataType::Date32, "timezone"));
-        assert!(temporal_accessor_valid(
-            &ScalarValue::Time64Nanosecond(None).data_type(),
-            "nanosecond"
-        ));
-        assert!(!temporal_accessor_valid(
-            &duration_scalar(None).data_type(),
-            "monthsOfYear"
-        ));
-        assert!(temporal_accessor_valid(
-            &localdatetime_scalar(None).data_type(),
-            "year"
-        ));
-        assert!(temporal_accessor_valid(
-            &datetime_scalar(None).data_type(),
-            "offsetSeconds"
-        ));
-        assert!(!temporal_accessor_valid(&DataType::Utf8, "year"));
-        assert!(!temporal_accessor_valid(&DataType::Int64, "day"));
-    }
-
-    #[test]
     fn scalar_to_ir_literal_round_trips_a_list() {
         // A homogeneous list now stores (#1006): scalar List → IrLiteral::List
         // and back, element-wise — including a list of typed temporals.
@@ -14339,86 +11445,6 @@ mod tests {
     }
 
     #[test]
-    fn duration_and_temporal_runtime_udfs_cover_each_value_family_and_nulls() {
-        let d1 = crate::temporal::DurationValue {
-            months: 1,
-            days: 2,
-            seconds: 3,
-            nanos: 750_000_000,
-        };
-        let d2 = crate::temporal::DurationValue {
-            months: 2,
-            days: 3,
-            seconds: 4,
-            nanos: 500_000_000,
-        };
-        let d1 = duration_scalar(Some(d1));
-        let d2 = duration_scalar(Some(d2));
-
-        let parsed = invoke_test_udf(
-            &CypherDurationParse::new(),
-            vec![ScalarValue::Utf8(Some("P1M2DT3.5S".into()))],
-        )
-        .unwrap();
-        assert!(duration_struct_parts(parsed.as_any().downcast_ref().unwrap(), 0).is_some());
-        let invalid = invoke_test_udf(
-            &CypherDurationParse::new(),
-            vec![ScalarValue::Utf8(Some("invalid".into()))],
-        )
-        .unwrap();
-        assert!(duration_struct_parts(invalid.as_any().downcast_ref().unwrap(), 0).is_none());
-
-        for sign in [1, -1] {
-            let out = invoke_test_udf(
-                &CypherDurationAdd::new(),
-                vec![d1.clone(), d2.clone(), ScalarValue::Int64(Some(sign))],
-            )
-            .unwrap();
-            assert!(duration_struct_parts(out.as_any().downcast_ref().unwrap(), 0).is_some());
-        }
-        for (factor, divide) in [(2.0, false), (2.0, true)] {
-            let out = invoke_test_udf(
-                &CypherDurationScale::new(),
-                vec![
-                    d1.clone(),
-                    ScalarValue::Float64(Some(factor)),
-                    ScalarValue::Boolean(Some(divide)),
-                ],
-            )
-            .unwrap();
-            assert!(duration_struct_parts(out.as_any().downcast_ref().unwrap(), 0).is_some());
-        }
-
-        let temporal_values = [
-            date_scalar(Some(20_000)),
-            ScalarValue::Time64Nanosecond(Some(10)),
-            time_scalar(Some((10, 3_600))),
-            localdatetime_scalar(Some((20_000, 10))),
-            datetime_scalar(Some((20_000, 10, 0, Some("UTC".into())))),
-        ];
-        for temporal in temporal_values {
-            for sign in [1, -1] {
-                let out = invoke_test_udf(
-                    &CypherTemporalArith::new(),
-                    vec![temporal.clone(), d1.clone(), ScalarValue::Int64(Some(sign))],
-                )
-                .unwrap();
-                assert_eq!(out.data_type(), &temporal.data_type());
-                assert!(!out.is_null(0));
-            }
-        }
-        assert!(
-            invoke_test_udf(
-                &CypherTemporalArith::new(),
-                vec![ScalarValue::Int64(Some(1)), d1, ScalarValue::Int64(Some(1))],
-            )
-            .unwrap_err()
-            .to_string()
-            .contains("not a temporal value")
-        );
-    }
-
-    #[test]
     fn exact_zero_large_list_legacy_variant_order_branches() {
         use datafusion::arrow::array::{Array, ArrayRef, Int64Array, LargeListArray};
         use datafusion::arrow::buffer::{NullBuffer, OffsetBuffer, ScalarBuffer};
@@ -14495,33 +11521,6 @@ mod tests {
                 .iter()
                 .any(|field| field.data_type() == &DataType::Utf8)
         );
-    }
-
-    #[test]
-    fn exact_zero_temporal_udf_metadata_contracts_are_total() {
-        fn check<U: ScalarUDFImpl + 'static>(udf: U) {
-            assert!((&udf as &dyn ScalarUDFImpl).is::<U>());
-            assert!(!udf.name().is_empty());
-            let _ = udf.signature();
-            assert!(udf.return_type(&[DataType::Null]).is_ok());
-        }
-
-        check(CypherDurationBetween::new());
-        check(CypherTemporalArith::new());
-        check(CypherDurationParse::new());
-        check(CypherDurationAdd::new());
-        check(CypherDurationScale::new());
-        check(CypherDateProject::new());
-        check(CypherLocalTimeProject::new());
-        check(CypherLocalTimeTruncate::new());
-        check(CypherLocalDateTimeProject::new());
-        check(CypherLocalDateTimeTruncate::new());
-        check(CypherTimeProject::new());
-        check(CypherTimeTruncate::new());
-        check(CypherDateTimeProject::new());
-        check(CypherDateTimeTruncate::new());
-        check(CypherToString::new());
-        check(CypherDateTruncate::new());
     }
 
     #[test]
